@@ -1,209 +1,212 @@
 {
   "name": "docs.github.com",
   "private": true,
-  "main": "server.js",
+  "repository": "https://github.com/github/docs",
+  "license": "(MIT AND CC-BY-4.0)",
   "author": {
-    "email": "opensource+docs@github.com",
     "name": "GitHub",
+    "email": "opensource+docs@github.com",
     "url": "https://github.com/github/docs"
   },
-  "license": "(MIT AND CC-BY-4.0)",
+  "type": "module",
+  "imports": {
+    "#src/*": {
+      "default": "./src/*"
+    }
+  },
+  "exports": "./server.js",
+  "scripts": {
+    "build": "next build",
+    "debug": "cross-env NODE_ENV=development ENABLED_LANGUAGES=en nodemon --inspect server.js",
+    "dev": "cross-env npm start",
+    "fixture-dev": "cross-env ROOT=tests/fixtures npm start",
+    "fixture-test": "cross-env ROOT=tests/fixtures npm test -- tests/rendering-fixtures",
+    "index-test-fixtures": "node src/search/scripts/index-elasticsearch.js -l en -l ja -V ghae -V dotcom --index-prefix tests -- src/search/tests/fixtures/search-indexes",
+    "lint": "eslint '**/*.{js,mjs,ts,tsx}'",
+    "lint-content": "node src/content-linter/scripts/lint-content.js",
+    "lint-translation": "cross-env NODE_OPTIONS=--experimental-vm-modules jest src/content-linter/tests/lint-files.js",
+    "openapi-docs": "node src/rest/docs.js",
+    "playwright-test": "playwright test --project=\"Google Chrome\"",
+    "prepare": "husky install",
+    "prettier": "prettier -w \"**/*.{ts,tsx,js,mjs,scss,yml,yaml}\"",
+    "prettier-check": "prettier -c \"**/*.{ts,tsx,js,mjs,scss,yml,yaml}\"",
+    "prevent-pushes-to-main": "node script/prevent-pushes-to-main.js",
+    "rest-dev": "node src/rest/scripts/update-files.js",
+    "show-action-deps": "echo 'Action Dependencies:' && rg '^[\\s|-]*(uses:.*)$' .github -I -N --no-heading -r '$1$2' | sort | uniq | cut -c 7-",
+    "prestart": "node script/cmp-files.js package-lock.json .installed.package-lock.json || npm install && cp package-lock.json .installed.package-lock.json",
+    "start": "cross-env NODE_ENV=development ENABLED_LANGUAGES=en nodemon server.js",
+    "start-all-languages": "cross-env NODE_ENV=development nodemon server.js",
+    "start-for-playwright": "cross-env ROOT=tests/fixtures TRANSLATIONS_FIXTURE_ROOT=tests/fixtures/translations ENABLED_LANGUAGES=en,ja NODE_ENV=test node server.js",
+    "sync-rest": "node src/rest/scripts/update-files.js",
+    "sync-search": "cross-env NODE_OPTIONS='--max_old_space_size=8192' start-server-and-test sync-search-server 4002 sync-search-indices",
+    "sync-search-ghes-release": "cross-env GHES_RELEASE=1 start-server-and-test sync-search-server 4002 sync-search-indices",
+    "sync-search-indices": "node src/search/scripts/sync-search-indices.js",
+    "sync-search-server": "cross-env NODE_ENV=production PORT=4002 MINIMAL_RENDER=true CHANGELOG_DISABLED=true node server.js",
+    "sync-webhooks": "src/rest/scripts/update-files.js -o webhooks",
+    "test": "cross-env NODE_OPTIONS='--max_old_space_size=4096 --experimental-vm-modules' jest --logHeapUsage",
+    "test-watch": "cross-env NODE_OPTIONS=--experimental-vm-modules jest --watch --notify --notifyMode=change --coverage",
+    "tsc": "tsc --noEmit"
+  },
+  "lint-staged": {
+    "*.{js,mjs,ts,tsx}": "eslint --cache --fix",
+    "*.{js,mjs,scss,ts,tsx,yml,yaml}": "prettier --write",
+    "{content,data}/**/*.md": "npm run lint-content -- --precommit --paths"
+  },
   "dependencies": {
-    "@babel/core": "^7.12.9",
-    "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@babel/plugin-syntax-class-properties": "^7.12.1",
-    "@babel/plugin-transform-modules-amd": "^7.12.1",
-    "@babel/plugin-transform-modules-commonjs": "^7.12.1",
-    "@babel/plugin-transform-react-jsx": "^7.12.7",
-    "@babel/plugin-transform-runtime": "^7.11.0",
-    "@babel/preset-env": "^7.12.7",
-    "@babel/preset-react": "^7.12.7",
-    "@babel/runtime": "^7.11.2",
-    "@graphql-inspector/core": "^2.3.0",
-    "@graphql-tools/load": "^6.2.5",
-    "@primer/css": "^16.2.0",
-    "@primer/octicons": "^13.0.0",
-    "ajv": "^6.11.0",
-    "algoliasearch": "^3.35.1",
-    "assert": "^2.0.0",
-    "babel-loader": "^8.1.0",
-    "babel-preset-env": "^1.7.0",
-    "browser-date-formatter": "^3.0.3",
-    "change-case": "^3.1.0",
-    "cheerio": "^1.0.0-rc.3",
-    "compression": "^1.7.4",
+    "@elastic/elasticsearch": "7.11.0",
+    "@github/failbot": "0.8.3",
+    "@primer/behaviors": "^1.3.5",
+    "@primer/css": "^21.0.7",
+    "@primer/octicons": "^19.5.0",
+    "@primer/octicons-react": "^19.5.0",
+    "@primer/react": "35.29.0",
+    "accept-language-parser": "^1.5.0",
+    "ajv": "^8.11.0",
+    "ajv-errors": "^3.0.0",
+    "ajv-formats": "^2.1.1",
+    "bottleneck": "2.19.5",
+    "boxen": "7.1.1",
+    "cheerio": "^1.0.0-rc.11",
+    "cheerio-to-text": "0.2.4",
+    "classnames": "^2.3.1",
     "connect-datadog": "0.0.9",
-    "connect-slashes": "^1.4.0",
-    "cookie-parser": "^1.4.5",
-    "copy-webpack-plugin": "^6.0.3",
-    "cors": "^2.8.5",
-    "cross-env": "^7.0.2",
-    "css-loader": "^4.3.0",
-    "csurf": "^1.11.0",
-    "directory-tree": "^2.2.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "express-rate-limit": "^5.1.3",
-    "express-timeout-handler": "^2.2.0",
-    "flat": "^5.0.0",
-    "github-slugger": "^1.2.1",
-    "got": "^9.6.0",
-    "gray-matter": "^4.0.2",
-    "hast-util-from-parse5": "^6.0.1",
-    "hast-util-parse-selector": "^2.2.5",
-    "hast-util-select": "^4.0.2",
-    "hast-util-to-string": "^1.0.4",
-    "hastscript": "^6.0.0",
-    "helmet": "^3.21.2",
-    "highlightjs-graphql": "^1.0.2",
-    "hot-shots": "^8.2.0",
-    "html-entities": "^1.2.1",
+    "cookie-parser": "^1.4.6",
+    "dayjs": "^1.11.3",
+    "dotenv": "^16.0.1",
+    "express": "^4.18.1",
+    "express-rate-limit": "7.0.0",
+    "express-timeout-handler": "^2.2.2",
+    "fastest-levenshtein": "1.0.16",
+    "file-type": "18.3.0",
+    "flat": "^6.0.1",
+    "github-slugger": "^2.0.0",
+    "glob": "10.0.0",
+    "got": "^13.0.0",
+    "gray-matter": "^4.0.3",
+    "hast-util-from-parse5": "^8.0.1",
+    "hast-util-to-string": "^2.0.0",
+    "hastscript": "^8.0.0",
+    "helmet": "^7.0.0",
+    "highlight.js": "11.7.0",
+    "highlightjs-curl": "^1.3.0",
+    "hot-shots": "^10.0.0",
+    "html-entities": "^2.3.3",
+    "http-proxy-middleware": "2.0.6",
     "imurmurhash": "^0.1.4",
-    "is-url": "^1.2.4",
-    "js-cookie": "^2.2.1",
-    "js-yaml": "^3.14.0",
-    "linkinator": "^2.13.6",
-    "liquidjs": "^9.22.1",
-    "lodash": "^4.17.20",
-    "lunr": "^2.3.9",
-    "lunr-languages": "^1.4.0",
-    "mdast-util-from-markdown": "^0.8.4",
-    "mini-css-extract-plugin": "^1.4.1",
-    "mkdirp": "^1.0.3",
-    "morgan": "^1.9.1",
-    "node-fetch": "^2.6.1",
-    "parse5": "^6.0.1",
+    "is-svg": "5.0.0",
+    "javascript-stringify": "^2.1.0",
+    "js-cookie": "^3.0.1",
+    "js-yaml": "^4.1.0",
+    "liquidjs": "^10.7.0",
+    "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21",
+    "lowdb": "6.0.0",
+    "lowlight": "3.0.0",
+    "mdast-util-from-markdown": "^2.0.0",
+    "mdast-util-to-hast": "^13.0.1",
+    "mdast-util-to-markdown": "2.0.0",
+    "mdast-util-to-string": "^4.0.0",
+    "morgan": "^1.10.0",
+    "next": "13.5.4",
+    "ora": "^7.0.1",
+    "parse5": "7.1.2",
     "port-used": "^2.0.8",
-    "rate-limit-redis": "^2.1.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "redis": "^3.1.1",
-    "redis-mock": "^0.56.3",
-    "rehype-autolink-headings": "^2.0.5",
-    "rehype-highlight": "^4.1.0",
-    "rehype-raw": "^4.0.2",
-    "rehype-slug": "^2.0.3",
-    "rehype-stringify": "^6.0.1",
-    "remark-code-extra": "^1.0.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "rehype-highlight": "^7.0.0",
+    "rehype-raw": "^7.0.0",
+    "rehype-slug": "^6.0.0",
+    "rehype-stringify": "^10.0.0",
     "remark-gemoji-to-emoji": "^1.1.0",
-    "remark-parse": "^7.0.2",
-    "remark-rehype": "^5.0.0",
-    "resolve-url-loader": "^3.1.2",
-    "revalidator": "^0.3.1",
-    "rimraf": "^3.0.0",
-    "rss-parser": "^3.12.0",
-    "sass": "^1.26.3",
-    "sass-loader": "^9.0.2",
-    "search-with-your-keyboard": "1.1.0",
-    "semver": "^5.7.1",
-    "slash": "^3.0.0",
+    "remark-gfm": "^4.0.0",
+    "remark-parse": "^11.0.0",
+    "remark-rehype": "^11.0.0",
+    "rss-parser": "^3.13.0",
+    "scroll-anchoring": "^0.1.0",
+    "semver": "^7.5.4",
+    "sharp": "0.32.1",
+    "slash": "^5.0.0",
     "strip-html-comments": "^1.0.0",
-    "style-loader": "^1.2.1",
-    "throng": "^5.0.0",
-    "unified": "^8.4.2",
-    "unist-util-visit": "^2.0.3",
-    "uuid": "^8.3.0",
-    "walk-sync": "^1.1.4",
-    "webpack": "^5.30.0",
-    "webpack-cli": "^4.6.0"
+    "styled-components": "^5.3.5",
+    "swr": "^2.2.2",
+    "unified": "^11.0.3",
+    "unist-util-visit": "^5.0.0",
+    "url-template": "^3.0.0",
+    "walk-sync": "^3.0.0"
   },
   "devDependencies": {
-    "@actions/core": "^1.2.6",
-    "@actions/github": "^4.0.0",
-    "@octokit/rest": "^16.43.2",
-    "async": "^3.2.0",
-    "await-sleep": "0.0.1",
-    "aws-sdk": "^2.610.0",
-    "babel-eslint": "^10.1.0",
-    "chalk": "^4.0.0",
-    "commander": "^6.2.0",
-    "count-array-values": "^1.2.1",
+    "@actions/core": "^1.10.0",
+    "@actions/github": "^6.0.0",
+    "@axe-core/playwright": "^4.7.3",
+    "@github/markdownlint-github": "^0.4.1",
+    "@graphql-inspector/core": "^5.0.0",
+    "@graphql-tools/load": "^8.0.0",
+    "@jest/globals": "29.7.0",
+    "@octokit/rest": "^20.0.2",
+    "@playwright/test": "1.38.1",
+    "@types/imurmurhash": "^0.1.1",
+    "@types/js-cookie": "^3.0.3",
+    "@types/lodash": "^4.14.198",
+    "@types/react": "18.2.21",
+    "@types/react-dom": "^18.2.7",
+    "@typescript-eslint/eslint-plugin": "6.7.0",
+    "@typescript-eslint/parser": "6.7.0",
+    "chalk": "^5.0.1",
+    "change-case": "^5.0.2",
+    "commander": "^11.0.0",
+    "cross-env": "^7.0.3",
     "csp-parse": "0.0.2",
-    "csv-parse": "^4.8.8",
-    "csv-parser": "^2.3.3",
-    "dedent": "^0.7.0",
-    "del": "^4.1.1",
-    "dependency-check": "^4.1.0",
-    "domwaiter": "^1.1.0",
-    "eslint": "^7.13.0",
-    "eslint-config-standard": "^16.0.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
+    "dedent": "^1.0.1",
+    "eslint": "8.49.0",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-config-standard": "17.1.0",
+    "eslint-import-resolver-typescript": "^3.6.0",
+    "eslint-plugin-import": "2.28.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-n": "16.1.0",
+    "eslint-plugin-primer-react": "^4.0.1",
+    "eslint-plugin-promise": "6.1.1",
     "event-to-promise": "^0.8.0",
-    "glob": "^7.1.6",
-    "graphql": "^14.5.8",
-    "heroku-client": "^3.1.0",
+    "graphql": "^16.8.1",
     "http-status-code": "^2.1.0",
-    "husky": "^4.2.1",
-    "image-size": "^0.7.4",
-    "japanese-characters": "^1.1.0",
-    "javascript-stringify": "^2.0.1",
-    "jest": "^26.6.3",
-    "jest-expect-message": "^1.0.2",
-    "jest-github-actions-reporter": "^1.0.2",
-    "jest-puppeteer": "^4.4.0",
-    "jest-silent-reporter": "^0.2.1",
+    "husky": "^8.0.3",
+    "jest": "29.7.0",
+    "jest-expect-message": "1.1.3",
+    "jest-fail-on-console": "^3.1.1",
     "jest-slow-test-reporter": "^1.0.0",
-    "jimp": "^0.16.1",
-    "make-promises-safe": "^5.1.0",
-    "mime": "^2.4.4",
-    "mock-express-response": "^0.2.2",
-    "mockdate": "^3.0.2",
-    "nock": "^13.0.4",
-    "nodemon": "^2.0.4",
-    "npm-merge-driver-install": "^2.0.0",
-    "object-hash": "^2.0.1",
-    "pa11y-ci": "^2.4.0",
-    "prettier": "^2.1.2",
-    "process": "^0.11.10",
-    "puppeteer": "^5.5.0",
-    "replace": "^1.2.0",
-    "robots-parser": "^2.1.1",
-    "start-server-and-test": "^1.12.0",
-    "strip-ansi": "^6.0.0",
-    "supertest": "^4.0.2",
-    "url-template": "^2.0.8",
-    "webpack-dev-middleware": "^4.1.0",
-    "website-scraper": "^4.2.0",
-    "xlsx-populate": "^1.21.0"
+    "json-schema-merge-allof": "^0.8.1",
+    "kill-port": "2.0.1",
+    "lint-staged": "^15.0.1",
+    "markdownlint": "^0.28.2",
+    "markdownlint-rule-helpers": "^0.19.0",
+    "markdownlint-rule-search-replace": "^1.2.0",
+    "mdast-util-gfm-table": "^2.0.0",
+    "micromark-extension-gfm-table": "^2.0.0",
+    "mkdirp": "^3.0.0",
+    "mockdate": "^3.0.5",
+    "nock": "^13.3.3",
+    "nodemon": "3.0.1",
+    "npm-merge-driver-install": "^3.0.0",
+    "nth-check": "2.1.1",
+    "prettier": "^3.0.3",
+    "rimraf": "^5.0.0",
+    "robots-parser": "^3.0.0",
+    "sass": "^1.52.3",
+    "start-server-and-test": "^2.0.0",
+    "ts-jest": "29.1.1",
+    "typescript": "^5.2.2",
+    "unist-util-remove": "^4.0.0",
+    "unist-util-visit-parents": "6.0.1",
+    "website-scraper": "^5.3.1"
   },
-  "scripts": {
-    "start": "cross-env NODE_ENV=development ENABLED_LANGUAGES='en,ja' nodemon server.js",
-    "dev": "npm start",
-    "debug": "cross-env NODE_ENV=development ENABLED_LANGUAGES='en,ja' nodemon --inspect server.js",
-    "rest-dev": "script/rest/update-files.js && npm run dev",
-    "build": "cross-env NODE_ENV=production npx webpack --mode production",
-    "start-all-languages": "cross-env NODE_ENV=development nodemon server.js",
-    "lint": "eslint --fix . && prettier -w \"**/*.{yml,yaml}\"",
-    "lint-translation": "TEST_TRANSLATION=true jest content/lint-files",
-    "test": "jest && eslint . && prettier -c \"**/*.{yml,yaml}\" && npm run check-deps",
-    "prebrowser-test": "npm run build",
-    "browser-test": "start-server-and-test browser-test-server 4001 browser-test-tests",
-    "browser-test-server": "cross-env NODE_ENV=production PORT=4001 node server.js",
-    "browser-test-tests": "cross-env BROWSER=1 jest tests/browser/browser.js",
-    "sync-search": "start-server-and-test sync-search-server 4002 sync-search-indices",
-    "sync-search-dry-run": "DRY_RUN=1 npm run sync-search",
-    "sync-search-server": "cross-env NODE_ENV=production PORT=4002 node server.js",
-    "sync-search-indices": "script/sync-search-indices.js",
-    "test-watch": "jest --watch --notify --notifyMode=change --coverage",
-    "check-deps": "node script/check-deps.js",
-    "prevent-pushes-to-main": "node script/prevent-pushes-to-main.js",
-    "pa11y-ci": "pa11y-ci",
-    "pa11y-test": "start-server-and-test browser-test-server 4001 pa11y-ci",
-    "link-check": "start-server-and-test link-check-server 4002 link-check-test",
-    "link-check-server": "cross-env NODE_ENV=development ENABLED_LANGUAGES='en' PORT=4002 node server.js",
-    "link-check-test": "cross-env node script/check-internal-links.js",
-    "heroku-postbuild": "node script/early-access/clone-for-build.js && npm run build"
+  "optionalDependencies": {
+    "esm": "^3.2.25"
   },
   "engines": {
-    "node": ">= 14.0.0"
+    "node": "^18 || ^20"
   },
-  "repository": "https://github.com/github/docs",
-  "husky": {
-    "hooks": {
-      "pre-commit": "node script/prevent-translation-commits.js",
-      "pre-push": "npm run prevent-pushes-to-main"
-    }
-  }
+  "cacheDirectories": [
+    "node_modules",
+    ".next/cache"
+  ]
 }
