@@ -10,93 +10,156 @@
 // content model, add the entry to this list, and ensure you loop in the
 // content and/or content strategy team for review.
 
-module.exports = [
-  '2fa',
+export default [
+  '2FA',
+  'Access management',
+  'Accounts',
   'Action development',
+  'Actions',
+  'Actions Runner Controller',
+  'Administrator',
+  'Advanced Security',
+  'Alerts',
   'Amazon ECS',
+  'Analytics',
   'Android',
   'Ant',
-  'Analytics',
   'API',
+  'Auditing',
+  'Authentication',
   'Azure App Service',
+  'Azure Kubernetes Service',
   'Azure Pipelines',
+  'Azure Static Web Apps',
+  'Backups',
+  'Billing',
+  'C/C++',
+  'C#',
+  'Cancellation',
   'CD',
   'CI',
   'CircleCI',
+  'CLI',
+  'Clustering',
+  'Code scanning',
+  'CodeQL',
+  'Codespaces',
+  'Collaboration',
+  'Community',
+  'Copilot',
   'Contacting sponsors',
   'Containers',
+  'CVEs',
+  'Dependabot',
+  'Dependencies',
+  'Dependency graph',
+  'Dependency review',
+  'Deployment',
+  'Desktop',
+  'Developer',
+  'Device verification',
+  'Discounts',
+  'Discussions',
   'Docker',
+  'Downgrades',
+  'Early access',
+  'Education',
+  'Enterprise',
+  'Errors',
   'Events',
+  'Forks',
   'Fundamentals',
+  'Git',
+  'GitHub Apps',
+  'GitHub Connect',
+  'GitHub search',
+  'GitHub',
   'GitLab',
   'Google Kubernetes Engine',
   'Gradle',
   'GraphQL',
+  'Health',
+  'High availability',
+  'Identity',
+  'Infrastructure',
+  'Integration',
   'iOS',
+  'Issues',
   'Java',
   'JavaScript',
   'Jenkins',
+  'Kotlin',
+  'Legal',
+  'LFS',
+  'Licensing',
+  'Logging',
+  'Maintenance',
+  'Marketplace',
   'Maven',
   'Migration',
+  'Mobile',
+  'Monitoring',
+  'Networking',
   'Node',
+  'Nonprofits',
+  'Notifications',
+  'OAuth apps',
   'Open Source',
   'Organizations',
+  'Packages',
   'Packaging',
-  'Powershell',
+  'Pages',
+  'Performance',
+  'Permissions',
+  'Policies',
+  'Policy',
+  'PowerShell',
+  'Pre-receive hooks',
   'Privacy',
+  'Profile',
+  'Profiles',
   'Project management',
+  'Projects',
   'Publishing',
+  'Pull requests',
   'Python',
+  'Receipts',
+  'Repositories',
+  'REST',
   'Ruby',
+  'SARIF',
+  'Secret scanning',
+  'Secret store',
+  'Security advisories',
+  'Security overview',
+  'Security policies',
+  'Security updates',
   'Security',
+  'Set up',
+  'Spending limits',
   'Sponsors payments',
   'Sponsors profile',
+  'Sponsors',
+  'SSH',
+  'SSO',
+  'Storage',
+  'Support',
   'Swift',
+  'Teams',
+  'Transcripts',
   'Travis CI',
+  'Trials',
+  'Troubleshooting',
+  'Upgrades',
   'User account',
+  'Usernames',
+  'Version updates',
+  'Visual Studio Code',
+  'Vulnerabilities',
   'Webhooks',
   'Workflows',
-  'access management',
-  'accounts',
-  'api', // replace this with API
-  'billing',
-  'cli',
-  'codespaces',
-  'community',
-  'desktop',
-  'device verification',
-  'early access',
-  'enterprise',
-  'events', // replace this with Events
-  'github',
-  'github apps',
-  'github search',
-  'identity',
-  'issues',
-  'jobs',
-  'legal',
-  'marketplace',
-  'mobile',
-  'notifications',
-  'oauth apps',
-  'open source', // replace this with Open Source
-  'organizations', // replace this with Organization
-  'pages',
-  'permissions',
-  'policy',
-  'profile',
-  'profiles',
-  'projects',
-  'pull requests',
-  'repositories',
-  'security',
-  'ssh',
-  'sso',
-  'teams',
-  'usernames',
-  'webhooks', // replace this with Webhooks
-  'Xamarin',
-  'Xamarin.iOS',
   'Xamarin.Android',
-  'Xcode'
+  'Xamarin.iOS',
+  'Xamarin',
+  'Xcode',
 ]
