@@ -1,26 +1,32 @@
-const fs = require('fs')
-const path = require('path')
-const frontmatter = require('./read-frontmatter')
-const getApplicableVersions = require('./get-applicable-versions')
-const removeFPTFromPath = require('./remove-fpt-from-path')
+import fs from 'fs/promises'
+import path from 'path'
+import frontmatter from './read-frontmatter.js'
+import getApplicableVersions from '#src/versions/lib/get-applicable-versions.js'
+import removeFPTFromPath from '#src/versions/lib/remove-fpt-from-path.js'
+import { ROOT } from './constants.js'
 
 // Both internal and external products are specified in content/index.md
-const homepage = path.posix.join(process.cwd(), 'content/index.md')
-const { data } = frontmatter(fs.readFileSync(homepage, 'utf8'))
-const productIds = data.children
-const externalProducts = data.externalProducts
+const homepage = path.posix.join(ROOT, 'content/index.md')
+export const { data } = frontmatter(await fs.readFile(homepage, 'utf8'))
 
+export const productIds = data.children
+
+const externalProducts = data.externalProducts
 const internalProducts = {}
 
-productIds.forEach(productId => {
+for (const productId of productIds) {
   const relPath = productId
-  const dir = path.posix.join('content', relPath)
+  const dir = path.join(ROOT, 'content', relPath)
 
   // Early Access may not exist in the current checkout
-  if (!fs.existsSync(dir)) return
+  try {
+    await fs.readdir(dir)
+  } catch (e) {
+    continue
+  }
 
   const toc = path.posix.join(dir, 'index.md')
-  const { data } = frontmatter(fs.readFileSync(toc, 'utf8'))
+  const { data } = frontmatter(await fs.readFile(toc, 'utf8'))
   const applicableVersions = getApplicableVersions(data.versions, toc)
   const href = removeFPTFromPath(path.posix.join('/', applicableVersions[0], productId))
 
@@ -31,15 +37,9 @@ productIds.forEach(productId => {
     dir,
     toc,
     wip: data.wip || false,
-    hidden: data.hidden || false
+    hidden: data.hidden || false,
+    versions: applicableVersions,
   }
-
-  internalProducts[productId].versions = applicableVersions
-})
-
-const productMap = Object.assign({}, internalProducts, externalProducts)
-
-module.exports = {
-  productIds,
-  productMap
 }
+
+export const productMap = Object.assign({}, internalProducts, externalProducts)
