@@ -1,5 +1,10 @@
-const express = require('express')
+import express from 'express'
+import middleware from '../middleware/index.js'
 
-const app = express()
-require('../middleware')(app)
-module.exports = app
+function createApp() {
+  const app = express()
+  middleware(app)
+  return app
+}
+
+export default createApp
