@@ -1,22 +1,19 @@
-const assert = require('assert')
-const fs = require('fs')
-const path = require('path')
-const walk = require('walk-sync')
-const yaml = require('js-yaml')
-const { isRegExp, set } = require('lodash')
-const filenameToKey = require('./filename-to-key')
+import assert from 'assert'
+import fs from 'fs'
+import path from 'path'
+import walk from 'walk-sync'
+import yaml from 'js-yaml'
+import { isRegExp, setWith } from 'lodash-es'
+import filenameToKey from './filename-to-key.js'
+import matter from 'gray-matter'
 
-module.exports = function dataDirectory (dir, opts = {}) {
+export default function dataDirectory(dir, opts = {}) {
   const defaultOpts = {
-    preprocess: (content) => { return content },
+    preprocess: (content) => {
+      return content
+    },
     ignorePatterns: [/README\.md$/i],
-    extensions: [
-      '.json',
-      '.md',
-      '.markdown',
-      '.yaml',
-      '.yml'
-    ]
+    extensions: ['.json', '.md', '.markdown', '.yml'],
   }
 
   opts = Object.assign({}, defaultOpts, opts)
@@ -31,18 +28,15 @@ module.exports = function dataDirectory (dir, opts = {}) {
   const data = {}
 
   // find YAML and Markdown files in the given directory, recursively
-  const filenames = walk(dir, { includeBasePath: true })
-    .filter(filename => {
-      // ignore files that match any of ignorePatterns regexes
-      if (opts.ignorePatterns.some(pattern => pattern.test(filename))) return false
+  const filenames = walk(dir, { includeBasePath: true }).filter((filename) => {
+    // ignore files that match any of ignorePatterns regexes
+    if (opts.ignorePatterns.some((pattern) => pattern.test(filename))) return false
 
-      // ignore files that don't have a whitelisted file extension
-      return opts.extensions.includes(path.extname(filename).toLowerCase())
-    })
+    // ignore files that don't have a whitelisted file extension
+    return opts.extensions.includes(path.extname(filename).toLowerCase())
+  })
 
-  const files = filenames.map(
-    filename => [filename, fs.readFileSync(filename, 'utf8')]
-  )
+  const files = filenames.map((filename) => [filename, fs.readFileSync(filename, 'utf8')])
   files.forEach(([filename, fileContent]) => {
     // derive `foo.bar.baz` object key from `foo/bar/baz.yml` filename
     const key = filenameToKey(path.relative(dir, filename))
@@ -50,22 +44,27 @@ module.exports = function dataDirectory (dir, opts = {}) {
 
     if (opts.preprocess) fileContent = opts.preprocess(fileContent)
 
-    // add this file's data to the global data object
+    // Add this file's data to the global data object.
+    // Note we want to use `setWith` instead of `set` so we can customize the type during path creation.
+    // If we just use `set`, then e.g. `release-notes.enterprise-server.2-20.0` will be an Array but
+    // `release-notes.enterprise-server.3-0.0` will be an Object.
+    // See https://lodash.com/docs#set for an explanation.
     switch (extension) {
       case '.json':
-        set(data, key, JSON.parse(fileContent))
+        setWith(data, key, JSON.parse(fileContent), Object)
         break
-      case '.yaml':
       case '.yml':
-        set(data, key, yaml.safeLoad(fileContent, { filename }))
+        setWith(data, key, yaml.load(fileContent, { filename }), Object)
         break
       case '.md':
       case '.markdown':
-        set(data, key, fileContent)
+        // Use `matter` to drop frontmatter, since localized reusable Markdown files
+        // can potentially have frontmatter, but we want to prevent the frontmatter
+        // from being rendered.
+        setWith(data, key, matter(fileContent).content, Object)
         break
     }
-  }
-  )
+  })
 
   return data
 }
