@@ -1,6 +1,6 @@
+import path from 'path'
+import { escapeRegExp } from 'lodash-es'
 /* eslint-disable prefer-regex-literals */
-const path = require('path')
-const { escapeRegExp } = require('lodash')
 
 // slash at the beginning of a filename
 const leadingPathSeparator = new RegExp(`^${escapeRegExp(path.sep)}`)
@@ -14,7 +14,7 @@ const windowsPathSeparator = new RegExp('/', 'g')
 const windowsDoubleSlashSeparator = new RegExp('\\\\', 'g')
 
 // derive `foo.bar.baz` object key from `foo/bar/baz.yml` filename
-module.exports = function filenameToKey (filename) {
+export default function filenameToKey(filename) {
   const extension = new RegExp(`${path.extname(filename)}$`)
   const key = filename
     .replace(extension, '')
