@@ -1,11 +1,17 @@
-const { getLanguageCode } = require('./patterns')
+import { getLanguageCode } from './patterns.js'
+import getRedirect from '#src/redirects/lib/get-redirect.js'
 
-module.exports = function findPage (href, pageMap, redirects) {
+export default function findPage(href, pages, redirects) {
+  if (Array.isArray(pages)) throw new Error("'pages' is not supposed to be an array")
+  if (pages === undefined) throw new Error("'pages' cannot be undefined")
+
   // remove any fragments
-  href = href.replace(/#.*$/, '')
+  href = new URL(href, 'http://example.com').pathname
+
+  const redirectsContext = { redirects, pages }
 
   // find the page
-  const page = pageMap[href] || pageMap[redirects[href]]
+  const page = pages[href] || pages[getRedirect(href, redirectsContext)]
   if (page) return page
 
   // get the current language
@@ -13,5 +19,5 @@ module.exports = function findPage (href, pageMap, redirects) {
 
   // try to fall back to English if the translated page can't be found
   const englishHref = href.replace(`/${currentLang}/`, '/en/')
-  return pageMap[englishHref] || pageMap[redirects[englishHref]]
+  return pages[englishHref] || pages[getRedirect(englishHref, redirectsContext)]
 }
