@@ -1,37 +1,56 @@
-// capture h2, h3, and h4 headings
-const headingRegex = /^###?#? (.*?)$/gm
+import { fromMarkdown } from 'mdast-util-from-markdown'
+import { toString } from 'mdast-util-to-string'
+import { visit } from 'unist-util-visit'
+
+import findPage from './find-page.js'
+import { getDataByLanguage } from './get-data.js'
 
 // for any translated page, first get corresponding English markdown
 // then get the headings on both the translated and English pageMap
 // finally, create a map of translation:English for all headings on the page
-module.exports = function getEnglishHeadings (page, context) {
+export default function getEnglishHeadings(page, context) {
   // Special handling for glossaries, because their headings are
   // generated programatically.
   if (page.relativePath.endsWith('/github-glossary.md')) {
     // Return an object of `{ localized-term: english-slug }`
-    const languageGlossary = context.site.data.glossaries.external
+    const languageGlossary = getDataByLanguage('glossaries.external', 'en')
     return languageGlossary.reduce((prev, curr) => {
       prev[curr.term] = curr.slug
       return prev
     }, {})
   }
 
-  const translatedHeadings = page.markdown.match(headingRegex)
-  if (!translatedHeadings) return
+  const translatedHeadings = getHeadings(page.markdown)
+  if (!translatedHeadings.length) return
 
-  const englishPage = context.pages[`/en/${page.relativePath.replace(/.md$/, '')}`]
+  const englishPage = findPage(
+    `/en/${page.relativePath.replace(/.md$/, '')}`,
+    context.pages,
+    context.redirects,
+  )
   if (!englishPage) return
 
-  // FIX  there may be bugs if English headings are updated before Crowdin syncs up :/
-  const englishHeadings = englishPage.markdown.match(headingRegex)
-  if (!englishHeadings) return
-
-  // select heading text only
-  const cleanTranslatedHeadings = translatedHeadings.map(h => h.replace(headingRegex, '$1'))
-  const cleanEnglishHeadings = englishHeadings.map(h => h.replace(headingRegex, '$1'))
+  // FIX  there may be bugs if English headings are updated before translations sync up :/
+  const englishHeadings = getHeadings(englishPage.markdown)
+  if (!englishHeadings.length) return
 
   // return a map from translation:English
-  return Object.assign(...cleanTranslatedHeadings.map((k, i) => ({
-    [k]: cleanEnglishHeadings[i]
-  })))
+  return Object.assign(
+    ...translatedHeadings.map((k, i) => ({
+      [k]: englishHeadings[i],
+    })),
+  )
+}
+
+function getHeadings(markdown) {
+  const ast = fromMarkdown(markdown)
+  const headings = []
+
+  visit(ast, (node) => {
+    if (node.type !== 'heading') return
+    if (![2, 3, 4].includes(node.depth)) return
+    headings.push(toString(node))
+  })
+
+  return headings
 }
