@@ -1,13 +1,13 @@
-const { productMap } = require('./all-products')
+import { productMap } from './all-products.js'
 const productTOCs = Object.values(productMap)
-  .filter(product => !product.external)
-  .map(product => product.toc.replace('content/', ''))
+  .filter((product) => !product.external)
+  .map((product) => product.toc.replace('content/', ''))
 
 const linkString = /{% [^}]*?link.*? \/(.*?) ?%}/m
 const linksArray = new RegExp(linkString.source, 'gm')
 
 // return an array of objects like { type: 'category|maptopic|article', href: 'path' }
-module.exports = function getTocItems (page) {
+export default function getTocItems(page) {
   // only process product and category tocs
   if (!page.relativePath.endsWith('index.md')) return
   if (page.relativePath === 'index.md') return
@@ -23,14 +23,16 @@ module.exports = function getTocItems (page) {
     return []
   }
 
-  return rawItems.map(item => {
+  return rawItems.map((item) => {
     const tocItem = {}
 
     // a product's toc items are always categories
     // whereas a category's toc items can be either maptopics or articles
     tocItem.type = productTOCs.includes(page.relativePath)
       ? 'category'
-      : item.includes('topic_') ? 'maptopic' : 'article'
+      : item.includes('topic_')
+      ? 'maptopic'
+      : 'article'
 
     tocItem.href = item.match(linkString)[1]
 
