@@ -1,33 +1,45 @@
-const assert = require('assert')
-const path = require('path')
-const cheerio = require('cheerio')
-const patterns = require('./patterns')
-const getMapTopicContent = require('./get-map-topic-content')
-const getApplicableVersions = require('./get-applicable-versions')
-const generateRedirectsForPermalinks = require('./redirects/permalinks')
-const getEnglishHeadings = require('./get-english-headings')
-const getTocItems = require('./get-toc-items')
-const pathUtils = require('./path-utils')
-const Permalink = require('./permalink')
-const languages = require('./languages')
-const renderContent = require('./render-content')
-const { renderReact } = require('./react/engine')
-const { productMap } = require('./all-products')
-const slash = require('slash')
-const statsd = require('./statsd')
-const readFileContents = require('./read-file-contents')
-const getLinkData = require('./get-link-data')
-const getDocumentType = require('./get-document-type')
-const union = require('lodash/union')
+import assert from 'assert'
+import path from 'path'
+import cheerio from 'cheerio'
+import getApplicableVersions from '#src/versions/lib/get-applicable-versions.js'
+import generateRedirectsForPermalinks from '#src/redirects/lib/permalinks.js'
+import getEnglishHeadings from './get-english-headings.js'
+import getTocItems from './get-toc-items.js'
+import Permalink from './permalink.js'
+import { renderContent } from '#src/content-render/index.js'
+import processLearningTracks from '#src/learning-track/lib/process-learning-tracks.js'
+import { productMap } from './all-products.js'
+import slash from 'slash'
+import readFileContents from './read-file-contents.js'
+import getLinkData from '#src/learning-track/lib/get-link-data.js'
+import getDocumentType from '#src/events/lib/get-document-type.js'
+import { union } from 'lodash-es'
+import { allTools } from '#src/tools/lib/all-tools.js'
+import { renderContentWithFallback } from './render-with-fallback.js'
+import { deprecated, supported } from '#src/versions/lib/enterprise-server-releases.js'
+import { allPlatforms } from '#src/tools/lib/all-platforms.js'
+
+// We're going to check a lot of pages' "ID" (the first part of
+// the relativePath) against `productMap` to make sure it's valid.
+// To avoid having to do `Object.keys(productMap).includes(id)`
+// every single time, we turn it into a Set once.
+const productMapKeysAsSet = new Set(Object.keys(productMap))
+
+export class FrontmatterErrorsError extends Error {
+  constructor(message, frontmatterErrors) {
+    super(message)
+    this.frontmatterErrors = frontmatterErrors
+  }
+}
 
 class Page {
-  static async init (opts) {
+  static async init(opts) {
     opts = await Page.read(opts)
     if (!opts) return
     return new Page(opts)
   }
 
-  static async read (opts) {
+  static async read(opts) {
     assert(opts.languageCode, 'languageCode is required')
     assert(opts.relativePath, 'relativePath is required')
     assert(opts.basePath, 'basePath is required')
@@ -38,15 +50,45 @@ class Page {
     // Per https://nodejs.org/api/fs.html#fs_fs_exists_path_callback
     // its better to read and handle errors than to check access/stats first
     try {
-      const { data, content, errors: frontmatterErrors } = await readFileContents(fullPath, opts.languageCode)
+      const { data, content, errors: frontmatterErrors } = await readFileContents(fullPath)
+
+      // The `|| ''` is for pages that are purely frontmatter.
+      // So the `content` property will be `undefined`.
+      let markdown = content || ''
+
+      // When the base path comes from the fixture, we make an exception.
+      // The enterprise-server version numbers are constantly evolving,
+      // but the fixtures are supposed to work in general for any version.
+      // That's why we allow special "macros" in the fixture content.
+      // If the fixture content looks like this:
+      //
+      //    {% ifversion ghes > __GHES_DEPRECATED__[0] and ghes < __GHES_SUPPORTED__[-2] %}
+      //
+      // it actually means the exact same thing as:
+      //
+      //    {% ifversion ghes > 3.5 and ghes < 3.7 %}
+      //
+      // ...**at the time**. The point is that these numbers change meaning
+      // where as notations like `__GHES_DEPRECATED__[3]`
+      // or `__GHES_SUPPORTED__[0]` are static.
+      if (opts.basePath.split(path.sep).includes('fixtures')) {
+        supported.forEach((version, i, arr) => {
+          markdown = markdown.replaceAll(`__GHES_SUPPORTED__[${i}]`, version)
+          markdown = markdown.replaceAll(`__GHES_SUPPORTED__[-${arr.length - i}]`, version)
+        })
+        deprecated.forEach((version, i, arr) => {
+          markdown = markdown.replaceAll(`__GHES_DEPRECATED__[${i}]`, version)
+          markdown = markdown.replaceAll(`__GHES_DEPRECATED__[-${arr.length - i}]`, version)
+        })
+      }
 
       return {
         ...opts,
         relativePath,
         fullPath,
         ...data,
-        markdown: content,
-        frontmatterErrors
+        markdown,
+        frontmatterErrors,
       }
     } catch (err) {
       if (err.code === 'ENOENT') return false
@@ -54,12 +96,15 @@ class Page {
     }
   }
 
-  constructor (opts) {
-    Object.assign(this, { ...opts })
-
-    if (this.frontmatterErrors.length) {
-      throw new Error(JSON.stringify(this.frontmatterErrors, null, 2))
+  constructor(opts) {
+    if (opts.frontmatterErrors && opts.frontmatterErrors.length) {
+      throw new FrontmatterErrorsError(
+        `${opts.frontmatterErrors.length} frontmatter errors trying to load ${opts.fullPath}`,
+        opts.frontmatterErrors,
+      )
     }
+    delete opts.frontmatterErrors
+    Object.assign(this, { ...opts })
 
     // Store raw data so we can cache parsed versions
     this.rawIntro = this.intro
@@ -69,13 +114,7 @@ class Page {
     this.rawPermissions = this.permissions
     this.rawLearningTracks = this.learningTracks
     this.rawIncludeGuides = this.includeGuides
-    this.raw_product_video = this.product_video
-
-    if (this.introLinks) {
-      this.introLinks.rawQuickstart = this.introLinks.quickstart
-      this.introLinks.rawReference = this.introLinks.reference
-      this.introLinks.rawOverview = this.introLinks.overview
-    }
+    this.rawIntroLinks = this.introLinks
 
     // Is this the Homepage or a Product, Category, Topic, or Article?
     this.documentType = getDocumentType(this.relativePath)
@@ -83,17 +122,32 @@ class Page {
     // Get array of versions that the page is available in for fast lookup
     this.applicableVersions = getApplicableVersions(this.versions, this.fullPath)
 
-    // a page should only be available in versions that its parent product is available in
-    const versionsParentProductIsNotAvailableIn = this.applicableVersions
-      // only the homepage will not have this.parentProduct
-      .filter(availableVersion => this.parentProduct && !this.parentProduct.versions.includes(availableVersion))
+    // Only check the parent product ID for English because if a top-level
+    // product is edited in English, it will fail for translations until
+    // the next translation pipeline PR gets a chance to catch up.
+    if (this.languageCode === 'en') {
+      // a page should only be available in versions that its parent product is available in
+      const versionsParentProductIsNotAvailableIn = this.applicableVersions
+        // only the homepage will not have this.parentProduct
+        .filter(
+          (availableVersion) =>
+            this.parentProduct && !this.parentProduct.versions.includes(availableVersion),
+        )
 
-    if (versionsParentProductIsNotAvailableIn.length) {
-      throw new Error(`\`versions\` frontmatter in ${this.fullPath} contains ${versionsParentProductIsNotAvailableIn}, which ${this.parentProduct.id} product is not available in!`)
+      if (versionsParentProductIsNotAvailableIn.length) {
+        throw new Error(
+          `\`versions\` frontmatter in ${this.fullPath} contains ${versionsParentProductIsNotAvailableIn}, which ${this.parentProduct.id} product is not available in!`,
+        )
+      }
     }
 
     // derive array of Permalink objects
-    this.permalinks = Permalink.derive(this.languageCode, this.relativePath, this.title, this.versions)
+    this.permalinks = Permalink.derive(
+      this.languageCode,
+      this.relativePath,
+      this.title,
+      this.applicableVersions,
+    )
 
     if (this.relativePath.endsWith('index.md')) {
       // get an array of linked items in product and category TOCs
@@ -101,27 +155,21 @@ class Page {
     }
 
     // if this is an article and it doesn't have showMiniToc = false, set mini TOC to true
-    if (!this.relativePath.endsWith('index.md') && !this.mapTopic) {
-      this.showMiniToc = this.showMiniToc === false
-        ? this.showMiniToc
-        : true
+    if (!this.relativePath.endsWith('index.md')) {
+      this.showMiniToc = this.showMiniToc === false ? this.showMiniToc : true
     }
 
-    // Instrument the `_render` method, so externally we call #render
-    // but it's wrapped in a timer that reports to Datadog
-    this.render = statsd.asyncTimer(this._render.bind(this), 'page.render')
+    this.render = this._render.bind(this)
 
     return this
   }
 
-  buildRedirects () {
-    // create backwards-compatible old paths for page permalinks and frontmatter redirects
-    this.redirects = generateRedirectsForPermalinks(this.permalinks, this.redirect_from)
-    return this.redirects
+  buildRedirects() {
+    return generateRedirectsForPermalinks(this.permalinks, this.redirect_from || [])
   }
 
   // Infer the parent product ID from the page's relative file path
-  get parentProductId () {
+  get parentProductId() {
     // Each page's top-level content directory matches its product ID
     const id = this.relativePath.split('/')[0]
 
@@ -130,101 +178,79 @@ class Page {
 
     // make sure the ID is valid
     if (process.env.NODE_ENV !== 'test') {
-      assert(
-        Object.keys(productMap).includes(id),
-        `page ${this.fullPath} has an invalid product ID: ${id}`
-      )
+      assert(productMapKeysAsSet.has(id), `page ${this.fullPath} has an invalid product ID: ${id}`)
     }
 
     return id
   }
 
-  get parentProduct () {
+  get parentProduct() {
     return productMap[this.parentProductId]
   }
 
-  async renderTitle (context, opts = {}) {
-    return this.shortTitle
+  async renderTitle(context, opts = { preferShort: true }) {
+    return opts.preferShort && this.shortTitle
       ? this.renderProp('shortTitle', context, opts)
       : this.renderProp('title', context, opts)
   }
 
-  async _render (context) {
+  async _render(context) {
     // use English IDs/anchors for translated headings, so links don't break (see #8572)
     if (this.languageCode !== 'en') {
       const englishHeadings = getEnglishHeadings(this, context)
       context.englishHeadings = englishHeadings
     }
 
-    this.intro = await renderContent(this.rawIntro, context)
-    this.introPlainText = await renderContent(this.rawIntro, context, { textOnly: true })
-    this.title = await renderContent(this.rawTitle, context, { textOnly: true, encodeEntities: true })
-    this.shortTitle = await renderContent(this.shortTitle, context, { textOnly: true, encodeEntities: true })
-    this.product_video = await renderContent(this.raw_product_video, context, { textOnly: true })
+    this.intro = await renderContentWithFallback(this, 'rawIntro', context)
+    this.introPlainText = await renderContentWithFallback(this, 'rawIntro', context, {
+      textOnly: true,
+    })
+    this.title = await renderContentWithFallback(this, 'rawTitle', context, {
+      textOnly: true,
+      throwIfEmpty: true,
+    })
 
-    if (this.introLinks) {
-      this.introLinks.quickstart = await renderContent(this.introLinks.rawQuickstart, context, { textOnly: true })
-      this.introLinks.reference = await renderContent(this.introLinks.rawReference, context, { textOnly: true })
-      this.introLinks.overview = await renderContent(this.introLinks.rawOverview, context, { textOnly: true })
+    const html = await renderContentWithFallback(this, 'markdown', context)
+
+    // Adding communityRedirect for Discussions, Sponsors, and Codespaces - request from Product
+    if (
+      this.parentProduct &&
+      (this.parentProduct.id === 'discussions' ||
+        this.parentProduct.id === 'sponsors' ||
+        this.parentProduct.id === 'codespaces')
+    ) {
+      this.communityRedirect = {
+        name: 'Provide GitHub Feedback',
+        href: `https://github.com/community/community/discussions/categories/${this.parentProduct.id}`,
+      }
     }
-
-    let markdown = this.mapTopic
-      // get the map topic child articles from the siteTree
-      ? getMapTopicContent(this.parentProduct.id, context.siteTree, context.currentLanguage, context.currentVersion, context.currentPath)
-      : this.markdown
-
-    // If the article is interactive parse the React!
-    if (this.interactive) {
-      // Search for the react code comments to find the react components
-      const reactComponents = markdown.match(/<!--react-->(.*?)<!--end-react-->/gs)
-
-      // Render each of the react components in the markdown
-      await Promise.all(reactComponents.map(async (reactComponent) => {
-        let componentStr = reactComponent
-
-        // Remove the React comment indicators
-        componentStr = componentStr.replace('<!--react-->\n', '').replace('<!--react-->', '')
-        componentStr = componentStr.replace('\n<!--end-react-->', '').replace('<!--end-react-->', '')
-
-        // Get the rendered component
-        const renderedComponent = await renderReact(componentStr)
-
-        // Replace the react component with the rendered markdown
-        markdown = markdown.replace(reactComponent, renderedComponent)
-      }))
-    }
-
-    context.relativePath = this.relativePath
-    const html = await renderContent(markdown, context)
 
     // product frontmatter may contain liquid
-    if (this.product) {
-      this.product = await renderContent(this.rawProduct, context)
+    if (this.rawProduct) {
+      this.product = await renderContentWithFallback(this, 'rawProduct', context)
     }
 
     // permissions frontmatter may contain liquid
-    if (this.permissions) {
-      this.permissions = await renderContent(this.rawPermissions, context)
+    if (this.rawPermissions) {
+      this.permissions = await renderContentWithFallback(this, 'rawPermissions', context)
     }
 
-    if (this.learningTracks) {
-      const learningTracks = []
-      for await (const rawTrackName of this.rawLearningTracks) {
-        // Track names in frontmatter may include Liquid conditionals
-        const renderedTrackName = await renderContent(rawTrackName, context, { textOnly: true, encodeEntities: true })
-        if (!renderedTrackName) continue
+    // Learning tracks may contain Liquid and need to have versioning processed.
+    if (this.rawLearningTracks) {
+      const { learningTracks } = await processLearningTracks(this.rawLearningTracks, context)
+      this.learningTracks = learningTracks
+    }
 
-        const track = context.site.data['learning-tracks'][context.currentProduct][renderedTrackName]
-        if (!track) continue
-        learningTracks.push({
-          trackName: renderedTrackName,
-          title: await renderContent(track.title, context, { textOnly: true, encodeEntities: true }),
-          description: await renderContent(track.description, context, { textOnly: true, encodeEntities: true }),
-          guides: await getLinkData(track.guides, context)
+    // introLinks may contain Liquid and need to have versioning processed.
+    if (this.rawIntroLinks) {
+      const introLinks = {}
+      for (const [rawKey, value] of Object.entries(this.rawIntroLinks)) {
+        introLinks[rawKey] = await renderContent(value, context, {
+          textOnly: true,
         })
       }
 
-      this.learningTracks = learningTracks
+      this.introLinks = introLinks
     }
 
     if (this.rawIncludeGuides) {
@@ -234,7 +260,9 @@ class Page {
         const { page } = guide
         guide.type = page.type
         if (page.topics) {
-          this.allTopics = union(this.allTopics, page.topics)
+          this.allTopics = union(this.allTopics, page.topics).sort((a, b) =>
+            a.localeCompare(b, page.languageCode),
+          )
           guide.topics = page.topics
         }
         delete guide.page
@@ -243,30 +271,47 @@ class Page {
     }
 
     // set a flag so layout knows whether to render a mac/windows/linux switcher element
-    this.includesPlatformSpecificContent = (
-      html.includes('extended-markdown mac') ||
-      html.includes('extended-markdown windows') ||
-      html.includes('extended-markdown linux')
-    )
+    // Remember, the values of platform is matched in
+    // the handleInvalidQuerystringValues shielding middleware.
+    this.detectedPlatforms = allPlatforms.filter((platform) => {
+      // This matches `ghd-tool mac` but not `ghd-tool macos`
+      // Whereas `html.includes('ghd-tool mac')` would match both.
+      const regex = new RegExp(`ghd-tool ${platform}\\b|platform-${platform}\\b`)
+      return regex.test(html)
+    })
+    this.includesPlatformSpecificContent = this.detectedPlatforms.length > 0
+
+    // set flags for webui, cli, etc switcher element
+    this.detectedTools = Object.keys(allTools).filter((tool) => {
+      // This matches `ghd-tool jetbrain` but not `ghd-tool jetbrain_beta`
+      // Whereas `html.includes('ghd-tool jetbrain')` would match both.
+      const regex = new RegExp(`ghd-tool ${tool}\\b|tool-${tool}\\b`)
+      return regex.test(html)
+    })
+
+    // pass the list of all possible tools around to components and utilities that will need it later on
+    this.allToolsParsed = allTools
+
+    this.includesToolSpecificContent = this.detectedTools.length > 0
 
     return html
   }
 
   // Allow other modules (like custom liquid tags) to make one-off requests
   // for a page's rendered properties like `title` and `intro`
-  async renderProp (propName, context, opts = { unwrap: false }) {
+  async renderProp(propName, context, opts = { unwrap: false }) {
     let prop
     if (propName === 'title') {
-      prop = this.rawTitle
+      prop = 'rawTitle'
     } else if (propName === 'shortTitle') {
-      prop = this.rawShortTitle || this.rawTitle // fall back to title
+      prop = this.rawShortTitle ? 'rawShortTitle' : 'rawTitle'
     } else if (propName === 'intro') {
-      prop = this.rawIntro
+      prop = 'rawIntro'
     } else {
-      prop = this[propName]
+      prop = propName
     }
 
-    const html = await renderContent(prop, context, opts)
+    const html = await renderContentWithFallback(this, prop, context, opts)
 
     if (!opts.unwrap) return html
 
@@ -278,23 +323,9 @@ class Page {
   // infer current page's corresponding homepage
   // /en/articles/foo                          -> /en
   // /en/enterprise/2.14/user/articles/foo     -> /en/enterprise/2.14/user
-  static getHomepage (requestPath) {
+  static getHomepage(requestPath) {
     return requestPath.replace(/\/articles.*/, '')
-  }
-
-  // given a page path, return an array of objects containing hrefs
-  // for that page in all languages
-  static getLanguageVariants (href) {
-    const suffix = pathUtils.getPathWithoutLanguage(href)
-    return Object.values(languages).map(({ name, code, hreflang }) => { // eslint-disable-line
-      return {
-        name,
-        code,
-        hreflang,
-        href: `/${code}${suffix}`.replace(patterns.trailingSlash, '$1')
-      }
-    })
   }
 }
 
-module.exports = Page
+export default Page
