@@ -1,27 +1,29 @@
-const slash = require('slash')
-const path = require('path')
-const patterns = require('./patterns')
-const { latest } = require('./enterprise-server-releases')
-const { productIds } = require('./all-products')
-const allVersions = require('./all-versions')
+import slash from 'slash'
+import path from 'path'
+import patterns from './patterns.js'
+import { latest } from '#src/versions/lib/enterprise-server-releases.js'
+import { productIds } from './all-products.js'
+import { allVersions } from '#src/versions/lib/all-versions.js'
+import nonEnterpriseDefaultVersion from '#src/versions/lib/non-enterprise-default-version.js'
 const supportedVersions = new Set(Object.keys(allVersions))
-const nonEnterpriseDefaultVersion = require('./non-enterprise-default-version')
 
 // Add the language to the given HREF
 // /articles/foo -> /en/articles/foo
-function getPathWithLanguage (href, languageCode) {
-  return slash(path.posix.join('/', languageCode, getPathWithoutLanguage(href)))
-    .replace(patterns.trailingSlash, '$1')
+export function getPathWithLanguage(href, languageCode) {
+  return slash(path.posix.join('/', languageCode, getPathWithoutLanguage(href))).replace(
+    patterns.trailingSlash,
+    '$1',
+  )
 }
 
 // Remove the language from the given HREF
 // /en/articles/foo -> /articles/foo
-function getPathWithoutLanguage (href) {
+export function getPathWithoutLanguage(href) {
   return slash(href.replace(patterns.hasLanguageCode, '/'))
 }
 
 // Remove the version segment from the path
-function getPathWithoutVersion (href) {
+export function getPathWithoutVersion(href) {
   const versionFromPath = getVersionStringFromPath(href)
 
   // If the derived version is not found in the list of all versions, just return the HREF
@@ -31,12 +33,13 @@ function getPathWithoutVersion (href) {
 }
 
 // Return the version segment in a path
-function getVersionStringFromPath (href) {
+export function getVersionStringFromPath(href, supportedOnly = false) {
   href = getPathWithoutLanguage(href)
 
-  // Return immediately if this is a link to the homepage
-  if (href === '/') {
-    return 'homepage'
+  // Some URLs don't ever have a version in the URL and it won't be found
+  // if you continue digging deeper. So exit early with the default on these.
+  if (['/', '/categories.json'].includes(href)) {
+    return nonEnterpriseDefaultVersion
   }
 
   // Get the first segment
@@ -58,9 +61,19 @@ function getVersionStringFromPath (href) {
   }
 
   // If it's just a plan with no @release (e.g., `enterprise-server`), return the latest release
-  const planObject = Object.values(allVersions).find(v => v.plan === versionFromPath)
+  const planObject = Object.values(allVersions).find((v) => v.plan === versionFromPath)
   if (planObject) {
     return allVersions[planObject.latestVersion].version
+  }
+
+  // If the caller of this function explicitly wants to know if the
+  // version part is *not* supported, they get back `undefined`.
+  // But this function is used in many other places where it potentially
+  // doesn't care if the version is supported.
+  // For example, in lib/redirects/permalinks.js it needs to know if the
+  // URL didn't contain a valid version.
+  if (supportedOnly) {
+    return
   }
 
   // Otherwise, return the first segment as-is, which may not be a real supported version,
@@ -69,14 +82,14 @@ function getVersionStringFromPath (href) {
 }
 
 // Return the corresponding object for the version segment in a path
-function getVersionObjectFromPath (href) {
+export function getVersionObjectFromPath(href) {
   const versionFromPath = getVersionStringFromPath(href)
 
   return allVersions[versionFromPath]
 }
 
 // Return the product segment from the path
-function getProductStringFromPath (href) {
+export function getProductStringFromPath(href) {
   href = getPathWithoutLanguage(href)
 
   if (href === '/') return 'homepage'
@@ -85,12 +98,14 @@ function getProductStringFromPath (href) {
 
   if (pathParts.includes('early-access')) return 'early-access'
 
-  return productIds.includes(pathParts[2])
-    ? pathParts[2]
-    : pathParts[1]
+  // For rest pages the currentProduct should be rest
+  // We use this to show SidebarRest, which is a different sidebar than the rest of the site
+  if (pathParts[1] === 'rest') return 'rest'
+
+  return productIds.includes(pathParts[2]) ? pathParts[2] : pathParts[1]
 }
 
-function getCategoryStringFromPath (href) {
+export function getCategoryStringFromPath(href) {
   href = getPathWithoutLanguage(href)
 
   if (href === '/') return null
@@ -99,19 +114,17 @@ function getCategoryStringFromPath (href) {
 
   if (pathParts.includes('early-access')) return null
 
-  const productIndex = productIds.includes(pathParts[2])
-    ? 2
-    : 1
+  const productIndex = productIds.includes(pathParts[2]) ? 2 : 1
 
   return pathParts[productIndex + 1]
 }
 
-module.exports = {
+export default {
   getPathWithLanguage,
   getPathWithoutLanguage,
   getPathWithoutVersion,
   getVersionStringFromPath,
   getVersionObjectFromPath,
   getProductStringFromPath,
-  getCategoryStringFromPath
+  getCategoryStringFromPath,
 }
