@@ -1,13 +1,34 @@
-const assert = require('assert')
-const path = require('path')
-const patterns = require('./patterns')
-const getApplicableVersions = require('./get-applicable-versions')
-const allVersions = require('./all-versions')
-const nonEnterpriseDefaultVersion = require('./non-enterprise-default-version')
-const removeFPTFromPath = require('./remove-fpt-from-path')
+import assert from 'assert'
+import path from 'path'
+import patterns from './patterns.js'
+import removeFPTFromPath from '#src/versions/lib/remove-fpt-from-path.js'
 
+/*
+This class creates the "permalinks" that power a page's different versions,
+where the source for the versions is a content file's frontmatter. The
+page.permalinks is an array of objects that looks like this:
+[
+  {
+    "languageCode": "en",
+    "pageVersion": "free-pro-team@latest",
+    "relativePath": "billing/managing-billing-for-your-github-account/index.md",
+    "title": "Managing billing for your GitHub account",
+    "hrefWithoutLanguage": "/billing/managing-billing-for-your-github-account",
+    "href": "/en/billing/managing-billing-for-your-github-account"
+  },
+  {
+    "languageCode": "en",
+    "pageVersion": "enterprise-cloud@latest",
+    "relativePath": "billing/managing-billing-for-your-github-account/index.md",
+    "title": "Managing billing for your GitHub account",
+    "hrefWithoutLanguage": "/enterprise-cloud@latest/billing/managing-billing-for-your-github-account",
+    "href": "/en/enterprise-cloud@latest/billing/managing-billing-for-your-github-account"
+  }
+  ... and so on for each of the content file's supported versions.
+]
+*/
 class Permalink {
-  constructor (languageCode, pageVersion, relativePath, title) {
+  constructor(languageCode, pageVersion, relativePath, title) {
     this.languageCode = languageCode
     this.pageVersion = pageVersion
     this.relativePath = relativePath
@@ -15,46 +36,38 @@ class Permalink {
 
     const permalinkSuffix = this.constructor.relativePathToSuffix(relativePath)
 
-    this.href = removeFPTFromPath(path.posix.join('/', languageCode, pageVersion, permalinkSuffix))
-      .replace(patterns.trailingSlash, '$1')
-
-    this.pageVersionTitle = allVersions[pageVersion].versionTitle
+    this.hrefWithoutLanguage = removeFPTFromPath(
+      path.posix.join('/', pageVersion, permalinkSuffix),
+    ).replace(patterns.trailingSlash, '$1')
+    this.href = `/${languageCode}${
+      this.hrefWithoutLanguage === '/' ? '' : this.hrefWithoutLanguage
+    }`
 
     return this
   }
 
-  static derive (languageCode, relativePath, title, frontmatterVersions) {
+  static derive(languageCode, relativePath, title, applicableVersions) {
     assert(relativePath, 'relativePath is required')
     assert(languageCode, 'languageCode is required')
 
-    const applicableVersions = getApplicableVersions(frontmatterVersions, path.join(languageCode, relativePath))
-    const permalinks = applicableVersions
-      // skip the Dotcom homepage here because a special homepage permalink is added below
-      .filter(pageVersion => !(pageVersion === nonEnterpriseDefaultVersion && relativePath === 'index.md'))
-      .map(pageVersion => {
-        return new Permalink(languageCode, pageVersion, relativePath, title)
-      })
-
-    // special permalink for homepage
-    if (relativePath === 'index.md') {
-      const homepagePermalink = {
-        ...permalinks[0],
-        href: '/' + languageCode,
-        pageVersion: 'homepage',
-        pageVersionTitle: permalinks[0].title,
-        homepage: true
-      }
-      permalinks.push(homepagePermalink)
-    }
+    const permalinks = applicableVersions.map((pageVersion) => {
+      return new Permalink(languageCode, pageVersion, relativePath, title)
+    })
 
     return permalinks
   }
 
-  static relativePathToSuffix (relativePath) {
-    return '/' + relativePath
-      .replace('index.md', '')
-      .replace('.md', '')
+  static relativePathToSuffix(relativePath) {
+    if (relativePath === 'index.md') return '/'
+    // When you turn `foo/bar.md`, which is a file path, into a URL pathname,
+    // you just need to chop off the `.md` suffix.
+    // For `foo/parent/index.md` you can't just chop off the `index.md`
+    // because it would leave a trailing `/`.
+    return `/${relativePath.replace(indexmdSuffixRegex, '').replace(mdSuffixRegex, '')}`
   }
 }
 
-module.exports = Permalink
+const indexmdSuffixRegex = new RegExp(`${path.sep}index\\.md$`)
+const mdSuffixRegex = /\.md$/
+
+export default Permalink
