@@ -1,11 +1,11 @@
-const enterpriseServerReleases = require('../lib/enterprise-server-releases')
+import enterpriseServerReleases from '#src/versions/lib/enterprise-server-releases.js'
 
 const productNames = {
-  dotcom: 'GitHub.com'
+  dotcom: 'GitHub.com',
 }
 
-enterpriseServerReleases.all.forEach(version => {
+enterpriseServerReleases.all.forEach((version) => {
   productNames[version] = `Enterprise Server ${version}`
 })
 
-module.exports = productNames
+export default productNames
