@@ -1,19 +1,15 @@
-const readFileAsync = require('./readfile-async')
-const encodeBracketedParentheses = require('./encode-bracketed-parentheses')
-const fm = require('./frontmatter')
+import fs from 'fs/promises'
+
+import encodeBracketedParentheses from './encode-bracketed-parentheses.js'
+import fm from './frontmatter.js'
 
 /**
  * Read only the frontmatter from  file
  */
-module.exports = async function fmfromf (filepath, languageCode) {
-  let fileContent = await readFileAsync(filepath, 'utf8')
+export default async function fmfromf(filepath) {
+  let fileContent = await fs.readFile(filepath, 'utf8')
 
   fileContent = encodeBracketedParentheses(fileContent)
-
-  // TODO remove this when crowdin-support issue 66 has been resolved
-  if (languageCode !== 'en' && fileContent.includes(': verdadero')) {
-    fileContent = fileContent.replace(': verdadero', ': true')
-  }
 
   return fm(fileContent, { filepath })
 }
