@@ -1,66 +1,83 @@
-const matter = require('gray-matter')
-const revalidator = require('revalidator')
-const { difference, intersection } = require('lodash')
+import matter from 'gray-matter'
+import Ajv from 'ajv'
+import addErrors from 'ajv-errors'
+import addFormats from 'ajv-formats'
+import semver from 'semver'
 
-function readFrontmatter (markdown, opts = { validateKeyNames: false, validateKeyOrder: false }) {
-  const schema = opts.schema || { properties: {} }
+const ajv = new Ajv({ allErrors: true, allowUnionTypes: true })
+ajv.addKeyword({
+  keyword: 'translatable',
+})
+ajv.addFormat('semver', {
+  validate: (x) => semver.validRange(x),
+})
+addErrors(ajv)
+addFormats(ajv)
+
+function readFrontmatter(markdown, opts = {}) {
+  const schema = opts.schema || { type: 'object', properties: {} }
   const filepath = opts.filepath || null
 
   let content, data
   let errors = []
 
   try {
-    ({ content, data } = matter(markdown))
+    ;({ content, data } = matter(markdown))
   } catch (e) {
     const defaultReason = 'invalid frontmatter entry'
 
     const reason = e.reason
-      // make this common error message a little easier to understand
-      ? e.reason.startsWith('can not read a block mapping entry;') ? defaultReason : e.reason
+      ? // make this common error message a little easier to understand
+        e.reason.startsWith('can not read a block mapping entry;')
+        ? defaultReason
+        : e.reason
       : defaultReason
 
     const error = {
       reason,
-      message: 'YML parsing error!'
+      message: 'YML parsing error!',
     }
 
     if (filepath) error.filepath = filepath
     errors.push(error)
+    console.warn(errors)
+
     return { errors }
   }
 
-  const allowedKeys = Object.keys(schema.properties)
-  const existingKeys = Object.keys(data)
-  const expectedKeys = intersection(allowedKeys, existingKeys)
+  const ajvValidate = ajv.compile(schema)
+  const valid = ajvValidate(data)
 
-  ;({ errors } = revalidator.validate(data, schema))
-
-  // add filepath property to each error object
-  if (errors.length && filepath) {
-    errors = errors.map(error => Object.assign(error, { filepath }))
+  if (!valid) {
+    errors = ajvValidate.errors
   }
 
-  // validate key names
-  if (opts.validateKeyNames) {
-    const invalidKeys = difference(existingKeys, allowedKeys)
-    invalidKeys.forEach(key => {
-      const error = {
-        property: key,
-        message: `not allowed. Allowed properties are: ${allowedKeys.join(', ')}`
-      }
-      if (filepath) error.filepath = filepath
-      errors.push(error)
+  // Combine the AJV-supplied `instancePath` and `params` into a more user-friendly frontmatter path.
+  // For example, given:
+  //   "instancePath": "/versions",
+  //   "params": { "additionalProperty": "ftp" }
+  // return:
+  //   property: 'versions.ftp'
+  //
+  // The purpose is to help users understand that the error is on the `fpt` key within the `versions` object.
+  // Note if the error is on a top-level FM property like `title`, the `instancePath` will be empty.
+  const cleanPropertyPath = (params, instancePath) => {
+    const mainProps = Object.values(params)[0]
+    if (!instancePath) return mainProps
+
+    const prefixProps = instancePath.replace('/', '').replace(/\//g, '.')
+    return typeof mainProps !== 'object' ? `${prefixProps}.${mainProps}` : prefixProps
+  }
+
+  if (!valid && filepath) {
+    errors = ajvValidate.errors.map((error) => {
+      const userFriendly = {}
+      userFriendly.property = cleanPropertyPath(error.params, error.instancePath)
+      userFriendly.message = error.message
+      userFriendly.reason = error.keyword
+      userFriendly.filepath = filepath
+      return userFriendly
     })
-  }
-
-  // validate key order
-  if (opts.validateKeyOrder && existingKeys.join('') !== expectedKeys.join('')) {
-    const error = {
-      property: 'keys',
-      message: `keys must be in order. Current: ${existingKeys.join(',')}; Expected: ${expectedKeys.join(',')}`
-    }
-    if (filepath) error.filepath = filepath
-    errors.push(error)
   }
 
   return { content, data, errors }
@@ -72,4 +89,4 @@ function readFrontmatter (markdown, opts = { validateKeyNames: false, validateKe
 // stringify('some string', {some: 'frontmatter'})
 readFrontmatter.stringify = matter.stringify
 
-module.exports = readFrontmatter
+export default readFrontmatter
