@@ -1,69 +1,50 @@
-const statsd = require('./statsd')
-const { loadPages, loadPageMap } = require('./pages')
-const loadRedirects = require('./redirects/precompile')
-const loadSiteData = require('./site-data')
-const loadSiteTree = require('./site-tree')
+import statsd from '#src/observability/lib/statsd.js'
+import { loadUnversionedTree, loadSiteTree, loadPages, loadPageMap } from './page-data.js'
+import loadRedirects from '#src/redirects/lib/precompile.js'
 
 // Instrument these functions so that
 // it's wrapped in a timer that reports to Datadog
 const dog = {
+  loadUnversionedTree: statsd.asyncTimer(loadUnversionedTree, 'load_unversioned_tree'),
+  loadSiteTree: statsd.asyncTimer(loadSiteTree, 'load_site_tree'),
   loadPages: statsd.asyncTimer(loadPages, 'load_pages'),
   loadPageMap: statsd.asyncTimer(loadPageMap, 'load_page_map'),
   loadRedirects: statsd.asyncTimer(loadRedirects, 'load_redirects'),
-  loadSiteData: statsd.timer(loadSiteData, 'load_site_data'),
-  loadSiteTree: statsd.asyncTimer(loadSiteTree, 'load_site_tree')
 }
 
-// For local caching
-let pageList, pageMap, site, redirects, siteTree
+// For multiple-triggered Promise sharing
+let promisedWarmServer
 
-function isFullyWarmed () {
-  // NOTE: Yes, `pageList` is specifically excluded here as it is transient data
-  const fullyWarmed = !!(pageMap && site && redirects && siteTree)
-  return fullyWarmed
-}
-
-function getWarmedCache () {
-  return {
-    pages: pageMap,
-    site,
-    redirects,
-    siteTree
-  }
-}
-
-async function warmServer () {
+async function warmServer() {
   const startTime = Date.now()
 
   if (process.env.NODE_ENV !== 'test') {
     console.log('Priming context information...')
   }
 
-  if (!pageList) {
-    pageList = await dog.loadPages()
-  }
+  const unversionedTree = await dog.loadUnversionedTree()
+  const siteTree = await dog.loadSiteTree(unversionedTree)
+  const pageList = await dog.loadPages(unversionedTree)
+  const pageMap = await dog.loadPageMap(pageList)
+  const redirects = await dog.loadRedirects(pageList)
 
-  if (!site) {
-    site = dog.loadSiteData()
-  }
+  statsd.gauge('memory_heap_used', process.memoryUsage().heapUsed, ['event:warm-server'])
 
-  if (!pageMap) {
-    pageMap = await dog.loadPageMap(pageList)
-  }
-
-  if (!redirects) {
-    redirects = await dog.loadRedirects(pageList, pageMap)
-  }
-
-  if (!siteTree) {
-    siteTree = await dog.loadSiteTree(pageMap, site, redirects)
-  }
+  statsd.gauge('num_redirects', Object.keys(redirects).length)
+  statsd.gauge('num_page_map', Object.keys(pageMap).length)
+  statsd.gauge('num_page_list', Object.keys(pageList).length)
 
   if (process.env.NODE_ENV !== 'test') {
     console.log(`Context primed in ${Date.now() - startTime} ms`)
   }
 
-  return getWarmedCache()
+  return {
+    pages: pageMap,
+    redirects,
+    unversionedTree,
+    siteTree,
+    pageList,
+  }
 }
 
 // Instrument the `warmServer` function so that
@@ -72,11 +53,12 @@ dog.warmServer = statsd.asyncTimer(warmServer, 'warm_server')
 
 // We only want statistics if the priming needs to occur, so let's wrap the
 // real method and return early [without statistics] whenever possible
-module.exports = async function warmServerWrapper () {
-  // Bail out early if everything is properly ready to use
-  if (isFullyWarmed()) {
-    return getWarmedCache()
+export default async function warmServerWrapper() {
+  // Handle receiving multiple calls to this method from multiple page requests
+  // by holding the in-progress Promise and returning it instead of allowing
+  // the server to actually load all of the files multiple times.
+  if (!promisedWarmServer) {
+    promisedWarmServer = dog.warmServer()
   }
-
-  return dog.warmServer()
+  return promisedWarmServer
 }
