@@ -1,13 +1,31 @@
-const { getRssFeed, getChangelogItems } = require('../../lib/changelog')
+import { getChangelogItems } from '../../lib/changelog.js'
+import getApplicableVersions from '#src/versions/lib/get-applicable-versions.js'
 
-module.exports = async function whatsNewChangelog (req, res, next) {
+export default async function whatsNewChangelog(req, res, next) {
   if (!req.context.page) return next()
   if (!req.context.page.changelog) return next()
-  req.context.changelogUrl = req.context.page.changelog.label === 'education'
-    ? 'https://github.blog/category/community/education'
-    : `https://github.blog/changelog/label/${req.context.page.changelog.label}`
+  const label = req.context.page.changelog.label.split(/\s+/g).join('')
 
-  const feed = await getRssFeed(req.context.changelogUrl)
-  req.context.whatsNewChangelog = await getChangelogItems(req.context.page.changelog.prefix, feed)
+  // If there is no `versions` prop in the changelog frontmatter, assume the changelog should display in all versions.
+  if (req.context.page.changelog.versions) {
+    const changelogVersions = getApplicableVersions(req.context.page.changelog.versions)
+
+    // If the current version is not included, do not display a changelog.
+    if (!changelogVersions.includes(req.context.currentVersion)) {
+      return next()
+    }
+  }
+
+  const labelUrls = {
+    education: 'https://github.blog/category/community/education',
+    enterprise: 'https://github.blog/category/enterprise/',
+  }
+
+  req.context.changelogUrl = labelUrls[label] || `https://github.blog/changelog/label/${label}`
+
+  req.context.whatsNewChangelog = await getChangelogItems(
+    req.context.page.changelog.prefix,
+    req.context.changelogUrl,
+  )
   return next()
 }
