@@ -1,4 +1,3 @@
-module.exports = require('cookie-parser')(
-  process.env.COOKIE_SECRET,
-  require('../lib/cookie-settings')
-)
+import cookieParser from 'cookie-parser'
+import cookieSettings from '../lib/cookie-settings.js'
+export default cookieParser(process.env.COOKIE_SECRET, cookieSettings)
