@@ -1,126 +1,126 @@
-const { get } = require('lodash')
-const { liquid } = require('../lib/render-content')
-const patterns = require('../lib/patterns')
-const layouts = require('../lib/layouts')
-const getMiniTocItems = require('../lib/get-mini-toc-items')
-const Page = require('../lib/page')
-const statsd = require('../lib/statsd')
-const RedisAccessor = require('../lib/redis-accessor')
-const { isConnectionDropped } = require('./halt-on-dropped-connection')
+import { get } from 'lodash-es'
 
-const { HEROKU_RELEASE_VERSION } = process.env
-const pageCacheDatabaseNumber = 1
-const pageCacheExpiration = 24 * 60 * 60 * 1000 // 24 hours
+import FailBot from '#src/observability/lib/failbot.js'
+import patterns from '../lib/patterns.js'
+import getMiniTocItems from '../lib/get-mini-toc-items.js'
+import { pathLanguagePrefixed } from '#src/languages/lib/languages.js'
+import statsd from '#src/observability/lib/statsd.js'
+import { allVersions } from '#src/versions/lib/all-versions.js'
+import { isConnectionDropped } from './halt-on-dropped-connection.js'
+import { nextApp, nextHandleRequest } from './next.js'
+import { defaultCacheControl } from './cache-control.js'
 
-const pageCache = new RedisAccessor({
-  databaseNumber: pageCacheDatabaseNumber,
-  prefix: (HEROKU_RELEASE_VERSION ? HEROKU_RELEASE_VERSION + ':' : '') + 'rp',
-  // Allow for graceful failures if a Redis SET operation fails
-  allowSetFailures: true,
-  // Allow for graceful failures if a Redis GET operation fails
-  allowGetFailures: true,
-  name: 'page-cache'
-})
+const STATSD_KEY_RENDER = 'middleware.render_page'
+const STATSD_KEY_404 = 'middleware.render_404'
 
-// a list of query params that *do* alter the rendered page, and therefore should be cached separately
-const cacheableQueries = ['learn']
+async function buildRenderedPage(req) {
+  const { context } = req
+  const { page } = context
+  const path = req.pagePath || req.path
 
-function addCsrf (req, text) {
-  return text.replace('$CSRFTOKEN$', req.csrfToken())
+  const pageRenderTimed = statsd.asyncTimer(page.render, STATSD_KEY_RENDER, [`path:${path}`])
+
+  return await pageRenderTimed(context)
 }
 
-module.exports = async function renderPage (req, res, next) {
-  const page = req.context.page
+async function buildMiniTocItems(req) {
+  const { context } = req
+  const { page } = context
+
+  // get mini TOC items on articles
+  if (!page.showMiniToc) {
+    return
+  }
+
+  return getMiniTocItems(context.renderedPage, '')
+}
+
+export default async function renderPage(req, res) {
+  const { context } = req
+
+  // This is a contextualizing the request so that when this `req` is
+  // ultimately passed into the `Error.getInitialProps` function,
+  // which NextJS executes at runtime on errors, so that we can
+  // from there send the error to Failbot.
+  req.FailBot = FailBot
+
+  const { page } = context
+  const path = req.pagePath || req.path
 
   // render a 404 page
   if (!page) {
-    if (process.env.NODE_ENV !== 'test' && req.context.redirectNotFound) {
-      console.error(`\nTried to redirect to ${req.context.redirectNotFound}, but that page was not found.\n`)
-    }
-    return res.status(404).send(
-      addCsrf(
-        req,
-        await liquid.parseAndRender(layouts['error-404'], req.context)
+    if (process.env.NODE_ENV !== 'test' && context.redirectNotFound) {
+      console.error(
+        `\nTried to redirect to ${context.redirectNotFound}, but that page was not found.\n`,
       )
-    )
+    }
+
+    if (!pathLanguagePrefixed(req.path)) {
+      defaultCacheControl(res)
+      return res.status(404).type('text').send('Not found')
+    }
+
+    // The rest is "unhandled" requests where we don't have the page
+    // but the URL looks like a real page.
+
+    statsd.increment(STATSD_KEY_404, 1, [
+      `url:${req.url}`,
+      `ip:${req.ip}`,
+      `path:${req.path}`,
+      `referer:${req.headers.referer || ''}`,
+    ])
+
+    return nextApp.render404(req, res)
   }
 
+  // Just finish fast without all the details like Content-Length
   if (req.method === 'HEAD') {
-    return res.status(200).end()
+    return res.status(200).send('')
   }
 
-  // Remove any query string (?...) and/or fragment identifier (#...)
-  const { pathname, searchParams } = new URL(req.originalUrl, 'https://docs.github.com')
+  // Updating the Last-Modified header for substantive changes on a page for engineering
+  // Docs Engineering Issue #945
+  if (page.effectiveDate) {
+    // Note that if a page has an invalidate `effectiveDate` string value,
+    // it would be caught prior to this usage and ultimately lead to
+    // 500 error.
+    res.setHeader('Last-Modified', new Date(page.effectiveDate).toUTCString())
+  }
 
-  for (const queryKey in req.query) {
-    if (!cacheableQueries.includes(queryKey)) {
-      searchParams.delete(queryKey)
+  // Stop processing if the connection was already dropped
+  if (isConnectionDropped(req, res)) return
+
+  req.context.renderedPage = await buildRenderedPage(req)
+  req.context.miniTocItems = await buildMiniTocItems(req)
+
+  // Stop processing if the connection was already dropped
+  if (isConnectionDropped(req, res)) return
+
+  // Create string for <title> tag
+  page.fullTitle = page.title
+
+  // add localized ` - GitHub Docs` suffix to <title> tag (except for the homepage)
+  if (!patterns.homepagePath.test(path)) {
+    if (
+      req.context.currentVersion === 'free-pro-team@latest' ||
+      !allVersions[req.context.currentVersion]
+    ) {
+      page.fullTitle += ' - ' + context.site.data.ui.header.github_docs
+    } else {
+      const { versionTitle } = allVersions[req.context.currentVersion]
+      page.fullTitle += ' - '
+      // Some plans don't have the word "GitHub" in them.
+      // E.g. "Enterprise Server 3.5"
+      // In those cases manually prefix the word "GitHub" before it.
+      if (!versionTitle.includes('GitHub')) {
+        page.fullTitle += 'GitHub '
+      }
+      page.fullTitle += versionTitle + ' Docs'
     }
   }
-  const originalUrl = pathname + ([...searchParams].length > 0 ? `?${searchParams}` : '')
-
-  // Serve from the cache if possible (skip during tests)
-  const isCacheable = !process.env.CI && process.env.NODE_ENV !== 'test' && req.method === 'GET'
 
   // Is the request for JSON debugging info?
   const isRequestingJsonForDebugging = 'json' in req.query && process.env.NODE_ENV !== 'production'
-
-  if (isCacheable && !isRequestingJsonForDebugging) {
-    // Stop processing if the connection was already dropped
-    if (isConnectionDropped(req, res)) return
-
-    const cachedHtml = await pageCache.get(originalUrl)
-    if (cachedHtml) {
-      // Stop processing if the connection was already dropped
-      if (isConnectionDropped(req, res)) return
-
-      console.log(`Serving from cached version of ${originalUrl}`)
-      statsd.increment('page.sent_from_cache')
-      return res.send(addCsrf(req, cachedHtml))
-    }
-  }
-
-  // add page context
-  const context = Object.assign({}, req.context, { page })
-
-  // collect URLs for variants of this page in all languages
-  context.page.languageVariants = Page.getLanguageVariants(req.path)
-
-  // Stop processing if the connection was already dropped
-  if (isConnectionDropped(req, res)) return
-
-  // render page
-  context.renderedPage = await page.render(context)
-
-  // Stop processing if the connection was already dropped
-  if (isConnectionDropped(req, res)) return
-
-  // get mini TOC items on articles
-  if (page.showMiniToc) {
-    context.miniTocItems = getMiniTocItems(context.renderedPage, page.miniTocMaxHeadingLevel)
-  }
-
-  // handle special-case prerendered GraphQL objects page
-  if (req.path.endsWith('graphql/reference/objects')) {
-    // concat the markdown source miniToc items and the prerendered miniToc items
-    context.miniTocItems = context.miniTocItems.concat(req.context.graphql.prerenderedObjectsForCurrentVersion.miniToc)
-    context.renderedPage = context.renderedPage + req.context.graphql.prerenderedObjectsForCurrentVersion.html
-  }
-
-  // handle special-case prerendered GraphQL input objects page
-  if (req.path.endsWith('graphql/reference/input-objects')) {
-    // concat the markdown source miniToc items and the prerendered miniToc items
-    context.miniTocItems = context.miniTocItems.concat(req.context.graphql.prerenderedInputObjectsForCurrentVersion.miniToc)
-    context.renderedPage = context.renderedPage + req.context.graphql.prerenderedInputObjectsForCurrentVersion.html
-  }
-
-  // Create string for <title> tag
-  context.page.fullTitle = context.page.title
-
-  // add localized ` - GitHub Docs` suffix to <title> tag (except for the homepage)
-  if (!patterns.homepagePath.test(req.path)) {
-    context.page.fullTitle = context.page.fullTitle + ' - ' + context.site.data.ui.header.github_docs
-  }
 
   // `?json` query param for debugging request context
   if (isRequestingJsonForDebugging) {
@@ -130,21 +130,14 @@ module.exports = async function renderPage (req, res, next) {
     } else {
       // dump all the keys: ?json
       return res.json({
-        message: 'The full context object is too big to display! Try one of the individual keys below, e.g. ?json=page. You can also access nested props like ?json=site.data.reusables',
-        keys: Object.keys(context)
+        message:
+          'The full context object is too big to display! Try one of the individual keys below, e.g. ?json=page. You can also access nested props like ?json=site.data.reusables',
+        keys: Object.keys(context),
       })
     }
   }
 
-  // currentLayout is added to the context object in middleware/contextualizers/layouts
-  const output = await liquid.parseAndRender(req.context.currentLayout, context)
+  defaultCacheControl(res)
 
-  // First, send the response so the user isn't waiting
-  // NOTE: Do NOT `return` here as we still need to cache the response afterward!
-  res.send(addCsrf(req, output))
-
-  // Finally, save output to cache for the next time around
-  if (isCacheable) {
-    await pageCache.set(originalUrl, output, { expireIn: pageCacheExpiration })
-  }
+  return nextHandleRequest(req, res)
 }
