@@ -1,12 +1,12 @@
-const timeout = require('express-timeout-handler')
+import timeout from 'express-timeout-handler'
 
-// Heroku router requests timeout after 30 seconds. We should stop them earlier!
-const maxRequestTimeout = parseInt(process.env.REQUEST_TIMEOUT, 10) || 10000
+import statsd from '#src/observability/lib/statsd.js'
+import { MAX_REQUEST_TIMEOUT } from '../lib/constants.js'
 
-module.exports = timeout.handler({
+export default timeout.handler({
   // Default timeout for all endpoints
-  // To override for a given router/endpoint, use `require('express-timeout-handler').set(...)`
-  timeout: maxRequestTimeout,
+  // To override for a given router/endpoint, use `xExpressTimeoutHandler.set(...)`
+  timeout: MAX_REQUEST_TIMEOUT,
 
   // IMPORTANT:
   // We cannot allow the middleware to disable the `res` object's methods like
@@ -14,6 +14,16 @@ module.exports = timeout.handler({
   disable: [],
 
   onTimeout: function (req, res, next) {
+    // The `req.pagePath` can come later so it's not guaranteed to always
+    // be present. It's added by the `handle-next-data-path.js` middleware
+    // we translates those "cryptic" `/_next/data/...` URLs from
+    // client-side routing.
+    const incrementTags = [`path:${req.pagePath || req.path}`]
+    if (req.context?.currentCategory) {
+      incrementTags.push(`product:${req.context.currentCategory}`)
+    }
+    statsd.increment('middleware.timeout', 1, incrementTags)
+
     // Create a custom timeout error
     const timeoutError = new Error('Request timed out')
     timeoutError.statusCode = 503
@@ -21,7 +31,7 @@ module.exports = timeout.handler({
 
     // Pass the error to our Express error handler for consolidated processing
     return next(timeoutError)
-  }
+  },
 
   // Can also set an `onDelayedResponse` property IF AND ONLY IF you allow for disabling methods
 })
