@@ -1,5 +1,9 @@
+#!/usr/bin/env node
+import dotenv from 'dotenv'
+import { Octokit } from '@octokit/rest'
+
 if (!process.env.GITHUB_TOKEN) {
-  require('dotenv').config()
+  dotenv.config()
 }
 
 // this module needs to work in development, production, and GitHub Actions
@@ -11,8 +15,8 @@ if (!process.env.GITHUB_TOKEN) {
 const apiToken = process.env.GITHUB_TOKEN
 
 // See https://github.com/octokit/rest.js/issues/1207
-module.exports = function github () {
-  return require('@octokit/rest')({
-    auth: `token ${apiToken}`
+export default function github() {
+  return new Octokit({
+    auth: `token ${apiToken}`,
   })
 }
