@@ -1,84 +1,73 @@
-const revalidator = require('revalidator')
-const schema = require('../helpers/schemas/site-tree-schema')
-const latestEnterpriseRelease = require('../../lib/enterprise-server-releases').latest
-const { getJSON } = require('../helpers/supertest')
-const flat = require('flat')
-const japaneseCharacters = require('japanese-characters')
-const nonEnterpriseDefaultVersion = require('../../lib/non-enterprise-default-version')
+import Ajv from 'ajv'
+import { jest } from '@jest/globals'
+import schema from '../helpers/schemas/site-tree-schema.js'
+import EnterpriseServerReleases from '#src/versions/lib/enterprise-server-releases.js'
+import { loadSiteTree } from '../../lib/page-data.js'
+import nonEnterpriseDefaultVersion from '#src/versions/lib/non-enterprise-default-version.js'
+import { formatAjvErrors } from '../helpers/schemas.js'
+
+const latestEnterpriseRelease = EnterpriseServerReleases.latest
+
+const ajv = new Ajv({ allErrors: true })
+const siteTreeValidate = ajv.compile(schema.childPage)
 
 describe('siteTree', () => {
   jest.setTimeout(3 * 60 * 1000)
 
-  let siteTree, flatTree
-  beforeAll(async (done) => {
-    siteTree = await getJSON('/en?json=siteTree')
-    flatTree = flat(siteTree)
-    done()
+  let siteTree
+  beforeAll(async () => {
+    siteTree = await loadSiteTree()
   })
 
   test('has language codes as top-level keys', () => {
     expect('en' in siteTree).toBe(true)
-    expect('ja' in siteTree).toBe(true)
   })
 
-  test('object order', () => {
-    expect(Object.keys(siteTree)[0]).toBe('en')
-    expect(Object.keys(siteTree.en)[0]).toBe(nonEnterpriseDefaultVersion)
-    expect(Object.keys(siteTree.en[nonEnterpriseDefaultVersion].products.github.categories)[0]).toBe('/en/github/getting-started-with-github')
-  })
-
-  test('object structure', () => {
-    expect(nonEnterpriseDefaultVersion in siteTree.en).toBe(true)
-    expect(`enterprise-server@${latestEnterpriseRelease}` in siteTree.en).toBe(true)
-    expect(flatTree[`en.${nonEnterpriseDefaultVersion}.products.github.href`]).toBe('/en/github')
-    expect(flatTree[`en.${nonEnterpriseDefaultVersion}.products.github.categories./en/github/getting-started-with-github.href`]).toBe('/en/github/getting-started-with-github')
+  test('object order and structure', () => {
+    expect(siteTree.en[nonEnterpriseDefaultVersion].childPages[1].href).toBe('/en/get-started')
+    expect(siteTree.en[nonEnterpriseDefaultVersion].childPages[1].childPages[0].href).toBe(
+      '/en/get-started/quickstart',
+    )
   })
 
   describe('localized titles', () => {
-    test('titles for categories', () => {
-      const japaneseTitle = flatTree[`ja.${nonEnterpriseDefaultVersion}.products.github.categories./ja/github/getting-started-with-github.title`]
-      expect(typeof japaneseTitle).toBe('string')
-      expect(japaneseCharacters.presentIn(japaneseTitle)).toBe(true)
+    test('articles that include site data in liquid templating', async () => {
+      const ghesLatest = `enterprise-server@${latestEnterpriseRelease}`
+      const ghesSiteTree = siteTree.en[ghesLatest]
 
-      const englishTitle = flatTree[`en.${nonEnterpriseDefaultVersion}.products.github.categories./en/github/getting-started-with-github.title`]
-      expect(typeof englishTitle).toBe('string')
-      expect(japaneseCharacters.presentIn(englishTitle)).toBe(false)
-    })
+      // Find a page in the tree that we know contains Liquid
+      // TODO: use new findPageInSiteTree helper when it's available
+      const pageWithDynamicTitle = ghesSiteTree.childPages
+        .find((child) => child.href === `/en/${ghesLatest}/admin`)
+        .childPages.find((child) => child.href === `/en/${ghesLatest}/admin/installation`)
 
-    test.todo('titles for maptopics')
-
-    test.todo('titles for articles')
-
-    test('articles that include site data in liquid templating', () => {
-      const pageWithDynamicTitle = siteTree.en[`enterprise-server@${latestEnterpriseRelease}`]
-        .products.admin
-        .categories[`/en/enterprise-server@${latestEnterpriseRelease}/admin/enterprise-support`]
-      // Source frontmatter from article:
-      // title: 'Working with {{ site.data.variables.contact.github_support }}'
-      expect(pageWithDynamicTitle.title).toEqual('Working with GitHub Support')
+      // Confirm the raw title contains Liquid
+      expect(pageWithDynamicTitle.page.title).toEqual(
+        'Installing {% data variables.product.prodname_enterprise %}',
+      )
     })
   })
 
   test('object validation', () => {
-    const products = Object.values(siteTree.en[nonEnterpriseDefaultVersion].products)
-    expect(products.length).toBeGreaterThan(0)
+    const childPages = siteTree.en[nonEnterpriseDefaultVersion].childPages
+    expect(childPages.length).toBeGreaterThan(0)
 
-    products.forEach(product => {
-      const { valid, errors } = revalidator.validate(product, schema.product)
-      const expectation = JSON.stringify(errors, null, 2)
-      expect(valid, expectation).toBe(true)
-
-      Object.values(product.categories || {}).forEach(category => {
-        const { valid, errors } = revalidator.validate(category, schema.category)
-        const expectation = JSON.stringify(errors, null, 2)
-        expect(valid, expectation).toBe(true)
-
-        Object.values(category.maptopics || {}).forEach(maptopic => {
-          const { valid, errors } = revalidator.validate(maptopic, schema.maptopic)
-          const expectation = JSON.stringify(errors, null, 2)
-          expect(valid, expectation).toBe(true)
-        })
-      })
-    })
+    validate(siteTree.en[nonEnterpriseDefaultVersion])
   })
 })
+
+function validate(currentPage) {
+  ;(currentPage.childPages || []).forEach((childPage) => {
+    const valid = siteTreeValidate(childPage)
+    let errors
+
+    if (!valid) {
+      errors = `file ${childPage.page.fullPath}: ${formatAjvErrors(siteTreeValidate.errors)}`
+    }
+
+    expect(valid, errors).toBe(true)
+
+    // Run recurisvely until we run out of child pages
+    validate(childPage)
+  })
+}
