@@ -1,83 +1,108 @@
 // match plan@release
-// e.g., free-pro-team@latest, enterprise-server@2.22
+// e.g., free-pro-team@latest, enterprise-server@3.0
 const planPattern = '^[a-z-]+'
 const releasePattern = '[a-z0-9-.]+'
 const delimiter = '@'
 const versionPattern = `${planPattern}${delimiter}${releasePattern}`
 
-module.exports = {
+export default {
+  type: 'object',
+  additionalProperties: false,
+  required: [
+    'version',
+    'versionTitle',
+    'latestVersion',
+    'currentRelease',
+    'planTitle',
+    'shortName',
+    'releases',
+    'latestRelease',
+    'openApiBaseName',
+    'openApiVersionName',
+    'miscBaseName',
+    'miscVersionName',
+    'apiVersions',
+    'latestApiVersion',
+  ],
   properties: {
     version: {
-      required: true,
       description: 'the version string',
       type: 'string',
-      pattern: versionPattern
+      pattern: versionPattern,
     },
     versionTitle: {
-      required: true,
       description: 'the version title',
-      type: 'string'
+      type: 'string',
     },
     latestVersion: {
-      required: true,
       description: 'the version name that includes the latest release',
       type: 'string',
-      pattern: versionPattern
+      pattern: versionPattern,
     },
     currentRelease: {
-      required: true,
       description: 'the release substring in the version string',
       type: 'string',
-      pattern: releasePattern
+      pattern: releasePattern,
     },
     plan: {
       description: 'the plan substring in the version string',
       type: 'string',
-      pattern: planPattern
+      pattern: planPattern,
     },
     planTitle: {
-      required: true,
       description: 'the plan title', // this is the same as the version title, sans numbered release
-      type: 'string'
+      type: 'string',
+    },
+    shortName: {
+      description: 'the short name for the version to be used in Liquid conditionals',
+      type: 'string',
     },
     releases: {
-      required: true,
       description: 'an array of all supported releases for the version',
-      type: 'array'
+      type: 'array',
     },
     latestRelease: {
-      required: true,
       description: 'the value of the latest release',
       type: 'string',
-      pattern: releasePattern
+      pattern: releasePattern,
+    },
+    internalLatestRelease: {
+      description:
+        'the value of "latest" if a plan uses semantic versioning internally while displaying @latest externally',
+      type: 'string',
     },
     hasNumberedReleases: {
-      description: 'boolean indicating whether the plan has numbered releases; if not, the release defalts to "latest"',
-      type: 'boolean'
+      description:
+        'boolean indicating whether the plan has numbered releases; if not, the release defalts to "latest"',
+      type: 'boolean',
     },
     nonEnterpriseDefault: {
       description: 'boolean indicating whether the plan is the default non-Enterprise version', // helper if the plan name changes
-      type: 'boolean'
+      type: 'boolean',
     },
     openApiBaseName: {
-      required: true,
       description: 'base name used to map an openAPI schema name to the current version',
-      type: 'string'
+      type: 'string',
     },
     openApiVersionName: {
-      required: true,
       description: 'final name used to map an openAPI schema name to the current version',
-      type: 'string'
+      type: 'string',
     },
     miscBaseName: {
-      required: true,
       description: 'base name used to map GraphQL and webhook schema names to the current version',
-      type: 'string'
+      type: 'string',
     },
     miscVersionName: {
-      required: true,
       description: 'final name used to map GraphQL and webhook schema names to the current version',
-      type: 'string'
-    }
-  }
+      type: 'string',
+    },
+    apiVersions: {
+      description: 'calendar date version for REST API versions',
+      type: 'array',
+    },
+    latestApiVersion: {
+      description: 'latest calendar date version for REST API versions',
+      type: 'string',
+    },
+  },
 }
