@@ -1,90 +1,93 @@
-const path = require('path')
-const { isPlainObject } = require('lodash')
-const supertest = require('supertest')
-const app = require('../../lib/app')
-const enterpriseServerReleases = require('../../lib/enterprise-server-releases')
-const nonEnterpriseDefaultVersion = require('../../lib/non-enterprise-default-version')
-const Page = require('../../lib/page')
-const { get } = require('../helpers/supertest')
+import { fileURLToPath } from 'url'
+import path from 'path'
+import { isPlainObject } from 'lodash-es'
+import { describe, expect, jest, test } from '@jest/globals'
+
+import enterpriseServerReleases, {
+  deprecatedWithFunctionalRedirects,
+} from '#src/versions/lib/enterprise-server-releases.js'
+import Page from '../../lib/page.js'
+import { get, head } from '../helpers/e2etest.js'
+import versionSatisfiesRange from '../../src/versions/lib/version-satisfies-range.js'
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
 describe('redirects', () => {
   jest.setTimeout(5 * 60 * 1000)
 
   let redirects
-  beforeAll(async (done) => {
+  beforeAll(async () => {
     const res = await get('/en?json=redirects')
-    redirects = JSON.parse(res.text)
-    done()
+    redirects = JSON.parse(res.body)
   })
 
-  test('page.redirects is an array', async () => {
+  test('page.buildRedirects() returns an array', async () => {
     const page = await Page.init({
-      relativePath: 'github/collaborating-with-issues-and-pull-requests/about-branches.md',
+      relativePath:
+        'pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-branches.md',
       basePath: path.join(__dirname, '../../content'),
-      languageCode: 'en'
+      languageCode: 'en',
     })
-    page.buildRedirects()
-    expect(isPlainObject(page.redirects)).toBe(true)
+    const pageRedirects = page.buildRedirects()
+    expect(isPlainObject(pageRedirects)).toBe(true)
   })
 
-  test('dotcom homepage page.redirects', async () => {
+  test('dotcom homepage page.buildRedirects()', async () => {
     const page = await Page.init({
-      relativePath: 'github/index.md',
+      relativePath: 'issues/index.md',
       basePath: path.join(__dirname, '../../content'),
-      languageCode: 'en'
+      languageCode: 'en',
     })
-    page.buildRedirects()
-    expect(page.redirects[`/en/${nonEnterpriseDefaultVersion}/github`]).toBe('/en/github')
-    expect(page.redirects['/articles']).toBe('/en/github')
-    expect(page.redirects['/en/articles']).toBe('/en/github')
-    expect(page.redirects[`/en/${nonEnterpriseDefaultVersion}/articles`]).toBe('/en/github')
-    expect(page.redirects['/common-issues-and-questions']).toBe('/en/github')
-    expect(page.redirects['/en/common-issues-and-questions']).toBe('/en/github')
-    expect(page.redirects[`/en/enterprise/${enterpriseServerReleases.latest}/user/articles`]).toBe(`/en/enterprise-server@${enterpriseServerReleases.latest}/github`)
-    expect(page.redirects[`/en/enterprise/${enterpriseServerReleases.latest}/user/common-issues-and-questions`]).toBe(`/en/enterprise-server@${enterpriseServerReleases.latest}/github`)
-  })
-
-  test('converts single `redirect_from` strings values into arrays', async () => {
-    const page = await Page.init({
-      relativePath: 'github/collaborating-with-issues-and-pull-requests/about-conversations-on-github.md',
-      basePath: path.join(__dirname, '../../content'),
-      languageCode: 'en'
-    })
-    page.buildRedirects()
-    const expected = '/en/github/collaborating-with-issues-and-pull-requests/about-conversations-on-github'
-    expect(page.redirects['/en/articles/about-discussions-in-issues-and-pull-requests']).toBe(expected)
+    const pageRedirects = page.buildRedirects()
+    expect(pageRedirects['/about-issues']).toBe('/issues')
+    expect(pageRedirects['/creating-an-issue']).toBe('/issues')
+    expect(
+      pageRedirects[`/enterprise-server@${enterpriseServerReleases.latest}/about-issues`],
+    ).toBe(`/enterprise-server@${enterpriseServerReleases.latest}/issues`)
+    expect(
+      pageRedirects[`/enterprise-server@${enterpriseServerReleases.latest}/creating-an-issue`],
+    ).toBe(`/enterprise-server@${enterpriseServerReleases.latest}/issues`)
   })
 
   describe('query params', () => {
     test('are preserved in redirected URLs', async () => {
       const res = await get('/enterprise/admin?query=pulls')
       expect(res.statusCode).toBe(301)
-      const expected = `/en/enterprise-server@${enterpriseServerReleases.latest}/admin?query=pulls`
+      const expected = `/en/enterprise-server@${enterpriseServerReleases.latest}/search?query=pulls`
       expect(res.headers.location).toBe(expected)
     })
 
     test('have q= converted to query=', async () => {
       const res = await get('/en/enterprise/admin?q=pulls')
       expect(res.statusCode).toBe(301)
-      const expected = `/en/enterprise-server@${enterpriseServerReleases.latest}/admin?query=pulls`
+      const expected = `/en/enterprise-server@${enterpriseServerReleases.latest}/search?query=pulls`
       expect(res.headers.location).toBe(expected)
     })
 
-    test('work with redirected search paths', async () => {
-      const res = await get('/en/enterprise/admin/search?utf8=%E2%9C%93&query=pulls')
+    test('convert q= to query= on search page', async () => {
+      const res = await get('/en/search?q=pulls')
       expect(res.statusCode).toBe(301)
-      const expected = `/en/enterprise-server@${enterpriseServerReleases.latest}/admin?utf8=%E2%9C%93&query=pulls`
+      const expected = '/en/search?query=pulls'
+      expect(res.headers.location).toBe(expected)
+    })
+
+    test('have faq= not converted to query=', async () => {
+      // Don't confuse `?faq=` for `?q=` just because they both start with `q=`
+      // Docs internal #21945
+      const res = await get('/en/enterprise/admin?faq=pulls')
+      expect(res.statusCode).toBe(301)
+      const expected = `/en/enterprise-server@${enterpriseServerReleases.latest}/admin?faq=pulls`
       expect(res.headers.location).toBe(expected)
     })
 
     test('do not work on other paths that include "search"', async () => {
-      const reqPath = `/en/enterprise-server@${enterpriseServerReleases.latest}/admin/configuration/enabling-unified-search-between-github-enterprise-server-and-githubcom`
+      const reqPath = `/en/enterprise-server@${enterpriseServerReleases.latest}/admin/configuration/configuring-github-connect/enabling-unified-search-for-your-enterprise`
       const res = await get(reqPath)
       expect(res.statusCode).toBe(200)
     })
 
-    test('work on deprecated versions', async () => {
-      const res = await get('/enterprise/2.12/admin/search?utf8=%E2%9C%93&q=pulls')
+    test('Do not redirect to search if on GraphQL Explorer "search"', async () => {
+      const res = await get('/en/graphql/overview/explorer?query=anything')
       expect(res.statusCode).toBe(200)
     })
   })
@@ -93,58 +96,90 @@ describe('redirects', () => {
     test('are absent from all redirected URLs', async () => {
       const keys = Object.keys(redirects)
       expect(keys.length).toBeGreaterThan(100)
-      expect(keys.every(key => !key.endsWith('/') || key === '/')).toBe(true)
+      expect(keys.every((key) => !key.endsWith('/') || key === '/')).toBe(true)
     })
 
     test('are absent from all destination URLs', async () => {
-      const values = Object.values(redirects)
+      const values = Object.entries(redirects)
+        .filter(([, to]) => !to.includes('://'))
+        .map(([from_]) => from_)
       expect(values.length).toBeGreaterThan(100)
-      expect(values.every(value => !value.endsWith('/'))).toBe(true)
+      expect(values.every((value) => !value.endsWith('/'))).toBe(true)
     })
 
     test('are redirected for HEAD requests (not just GET requests)', async () => {
-      const res = await supertest(app).head('/articles/closing-issues-via-commit-messages/')
+      const res = await head('/articles/closing-issues-via-commit-messages/')
       expect(res.statusCode).toBe(301)
       expect(res.headers.location).toBe('/articles/closing-issues-via-commit-messages')
     })
   })
 
-  describe('external redirects', () => {
-    test('work for top-level request paths', async () => {
-      const res = await get('/git-ready')
-      expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe('http://gitready.com/')
+  describe('home page redirects', () => {
+    test('homepage redirects to english by default', async () => {
+      const res = await get('/')
+      expect(res.statusCode).toBe(302)
+      expect(res.headers.location).toBe('/en')
+      // language specific caching
+      expect(res.headers['cache-control']).toContain('public')
+      expect(res.headers['cache-control']).toMatch(/max-age=\d+/)
+      expect(res.headers.vary).toContain('accept-language')
+      expect(res.headers.vary).toContain('x-user-language')
     })
 
-    test('work for article-level request paths', async () => {
-      const res = await get('/articles/testing-webhooks')
+    test('trailing slash on languaged homepage should permanently redirect', async () => {
+      const res = await get('/en/')
       expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe('http://developer.github.com/webhooks/testing/')
+      expect(res.headers.location).toBe('/en')
+      expect(res.headers['set-cookie']).toBeUndefined()
+      expect(res.headers['cache-control']).toContain('public')
+      expect(res.headers['cache-control']).toMatch(/max-age=\d+/)
     })
   })
 
-  describe('localized redirects', () => {
-    test('redirect_from for renamed pages', async () => {
-      const { res } = await get('/ja/desktop/contributing-to-projects/changing-a-remote-s-url-from-github-desktop')
+  describe('external redirects', () => {
+    test('no external redirect starts with a language prefix', () => {
+      const values = Object.entries(redirects)
+        .filter(([, to]) => to.includes('://'))
+        .map(([from_]) => from_)
+        .filter((from_) => from_.startsWith('/en/'))
+      expect(values.length).toBe(0)
+    })
+
+    test('no external redirect should go to developer.github.com', () => {
+      const values = Object.values(redirects)
+        .filter((to) => to.includes('://'))
+        .filter((to) => new URL(to).hostname === 'developer.github.com')
+      expect(values.length).toBe(0)
+    })
+
+    test('work for top-level request paths', async () => {
+      const res = await get('/git-ready')
       expect(res.statusCode).toBe(301)
-      const expected = '/ja/desktop/contributing-and-collaborating-using-github-desktop/changing-a-remotes-url-from-github-desktop'
-      expect(res.headers.location).toBe(expected)
+      expect(res.headers.location).toBe('https://gitready.com/')
+      expect(res.headers['set-cookie']).toBeUndefined()
+      expect(res.headers['cache-control']).toContain('public')
+      expect(res.headers['cache-control']).toMatch(/max-age=\d+/)
+    })
+
+    test('work for top-level request paths with /en/ prefix', async () => {
+      const res = await get('/en/git-ready')
+      expect(res.statusCode).toBe(301)
+      expect(res.headers.location).toBe('https://gitready.com/')
     })
   })
 
   describe('enterprise home page', () => {
     const enterpriseHome = `/en/enterprise-server@${enterpriseServerReleases.latest}`
-    const japaneseEnterpriseHome = enterpriseHome.replace('/en/', '/ja/')
 
     test('/enterprise', async () => {
       const res = await get('/enterprise')
-      expect(res.statusCode).toBe(301)
+      expect(res.statusCode).toBe(302)
       expect(res.headers.location).toBe(enterpriseHome)
     })
 
     test('no language code redirects to english', async () => {
       const res = await get(`/enterprise/${enterpriseServerReleases.latest}`)
-      expect(res.statusCode).toBe(301)
+      expect(res.statusCode).toBe(302)
       expect(res.headers.location).toBe(enterpriseHome)
     })
 
@@ -154,15 +189,9 @@ describe('redirects', () => {
       expect(res.headers.location).toBe(enterpriseHome)
     })
 
-    test('no version redirects to latest version (japanese)', async () => {
-      const res = await get('/ja/enterprise')
-      expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe(japaneseEnterpriseHome)
-    })
-
     test('hardcoded @latest redirects to latest version', async () => {
       const res = await get('/en/enterprise-server@latest')
-      expect(res.statusCode).toBe(301)
+      expect(res.statusCode).toBe(302)
       expect(res.headers.location).toBe(enterpriseHome)
     })
   })
@@ -170,7 +199,7 @@ describe('redirects', () => {
   describe('2.13+ deprecated enterprise', () => {
     test('no language code redirects to english', async () => {
       const res = await get('/enterprise/2.13')
-      expect(res.statusCode).toBe(301)
+      expect(res.statusCode).toBe(302)
       expect(res.headers.location).toBe('/en/enterprise/2.13')
     })
 
@@ -190,18 +219,22 @@ describe('redirects', () => {
 
     test('frontmatter redirect', async () => {
       const res = await get('/enterprise/2.12/user/articles/github-flavored-markdown')
-      expect(res.statusCode).toBe(301)
-      expect(res.text).toContain('location=\'/enterprise/2.12/user/categories/writing-on-github/\'')
+      expect(res.statusCode).toBe(302) // because it doesn't have a language
+      expect(res.headers.location).toBe('/enterprise/2.12/user/categories/writing-on-github/')
     })
   })
 
   describe('enterprise admin', () => {
-    const enterpriseAdmin = `/en/enterprise-server@${enterpriseServerReleases.latest}/admin`
-    const japaneseEnterpriseAdmin = enterpriseAdmin.replace('/en/', '/ja/')
+    // firstRestoredAdminGuides = 2.21
+    // lastBeforeRestoredAdminGuides = 2.20
+    // (these won't change but it's more convenient to use constants)
+    const { firstRestoredAdminGuides, getPreviousReleaseNumber, latest } = enterpriseServerReleases
+    const lastBeforeRestoredAdminGuides = getPreviousReleaseNumber(firstRestoredAdminGuides)
+    const enterpriseAdmin = `/en/enterprise-server@${latest}/admin`
 
     test('no language code redirects to english', async () => {
-      const res = await get(`/enterprise/${enterpriseServerReleases.latest}/admin`)
-      expect(res.statusCode).toBe(301)
+      const res = await get(`/enterprise/${latest}/admin`)
+      expect(res.statusCode).toBe(302)
       expect(res.headers.location).toBe(enterpriseAdmin)
     })
 
@@ -211,43 +244,51 @@ describe('redirects', () => {
       expect(res.headers.location).toBe(enterpriseAdmin)
     })
 
-    test('admin/guides redirects to admin', async () => {
-      const res = await get(`/en/enterprise/${enterpriseServerReleases.latest}/admin/guides`)
+    test('admin/guides redirects to admin on <2.21', async () => {
+      const res = await get(`/en/enterprise-server@${lastBeforeRestoredAdminGuides}/admin/guides`)
       expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe(enterpriseAdmin)
+      expect(res.headers.location).toBe(
+        enterpriseAdmin.replace(latest, lastBeforeRestoredAdminGuides),
+      )
     })
 
-    test('no version plus admin/guides redirects to admin on latest version', async () => {
-      const res = await get('/en/enterprise/admin/guides')
-      expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe(enterpriseAdmin)
+    test('admin/guides does not redirect to admin on >=2.21', async () => {
+      const res = await get(`/en/enterprise-server@${firstRestoredAdminGuides}/admin/guides`)
+      expect(res.statusCode).toBe(200)
     })
 
-    test('admin/guides redirects to admin in redirects', async () => {
-      const res = await get(`/en/enterprise/${enterpriseServerReleases.latest}/admin/guides/installation/upgrading-github-enterprise`)
+    test('no version plus admin/guides redirects to the right place on latest version', async () => {
+      const shouldRedirect = versionSatisfiesRange(latest, `<${firstRestoredAdminGuides}`)
+      const expectedStatusCode = shouldRedirect ? 301 : 200
+      const res = await get(`/en/enterprise-server@${latest}/admin/guides`)
+      expect(res.statusCode).toBe(expectedStatusCode)
+    })
+
+    test('admin/guides redirects to admin in deep links on <2.21', async () => {
+      const res = await get(
+        `/en/enterprise-server@${lastBeforeRestoredAdminGuides}/admin/guides/installation/upgrading-github-enterprise`,
+      )
       expect(res.statusCode).toBe(301)
       const redirectRes = await get(res.headers.location)
       expect(redirectRes.statusCode).toBe(200)
-      const expected = `/en/enterprise-server@${enterpriseServerReleases.latest}/admin/enterprise-management/upgrading-github-enterprise-server`
+      const expected = `/en/enterprise-server@${lastBeforeRestoredAdminGuides}/admin/enterprise-management/upgrading-github-enterprise-server`
       expect(res.headers.location).toBe(expected)
     })
 
-    test('no version redirects to latest version (japanese)', async () => {
-      const res = await get('/ja/enterprise/admin')
+    test('admin/guides still redirects to admin in deep links on >=2.21', async () => {
+      const res = await get(
+        `/en/enterprise-server@${firstRestoredAdminGuides}/admin/guides/installation/upgrading-github-enterprise`,
+      )
       expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe(japaneseEnterpriseAdmin)
-    })
-
-    test('admin/guides redirects to admin (japanese)', async () => {
-      const res = await get(`/ja/enterprise/${enterpriseServerReleases.latest}/admin/guides`)
-      expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe(japaneseEnterpriseAdmin)
+      const redirectRes = await get(res.headers.location)
+      expect(redirectRes.statusCode).toBe(200)
+      const expected = `/en/enterprise-server@${firstRestoredAdminGuides}/admin/enterprise-management/updating-the-virtual-machine-and-physical-resources/upgrading-github-enterprise-server`
+      expect(res.headers.location).toBe(expected)
     })
   })
 
   describe('enterprise user homepage', () => {
-    const enterpriseUser = `/en/enterprise-server@${enterpriseServerReleases.latest}/github`
-    const japaneseEnterpriseUser = enterpriseUser.replace('/en/', '/ja/')
+    const enterpriseUser = `/en/enterprise-server@${enterpriseServerReleases.latest}`
 
     test('no product redirects to GitHub.com product', async () => {
       const res = await get(`/en/enterprise/${enterpriseServerReleases.latest}/user`)
@@ -256,8 +297,8 @@ describe('redirects', () => {
     })
 
     test('no language code redirects to english', async () => {
-      const res = await get(`/enterprise/${enterpriseServerReleases.latest}/user/github`)
-      expect(res.statusCode).toBe(301)
+      const res = await get(`/enterprise/${enterpriseServerReleases.latest}/user`)
+      expect(res.statusCode).toBe(302)
       expect(res.headers.location).toBe(enterpriseUser)
     })
 
@@ -266,20 +307,15 @@ describe('redirects', () => {
       expect(res.statusCode).toBe(301)
       expect(res.headers.location).toBe(enterpriseUser)
     })
-
-    test('no version redirects to latest version (japanese)', async () => {
-      const res = await get('/ja/enterprise/user/github')
-      expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe(japaneseEnterpriseUser)
-    })
   })
 
   describe('enterprise user article', () => {
-    const userArticle = `/en/enterprise-server@${enterpriseServerReleases.latest}/github/getting-started-with-github/set-up-git`
-    const japaneseUserArticle = userArticle.replace('/en/', '/ja/')
+    const userArticle = `/en/enterprise-server@${enterpriseServerReleases.latest}/get-started/quickstart/fork-a-repo`
 
     test('no product redirects to GitHub.com product on the latest version', async () => {
-      const res = await get(`/en/enterprise/${enterpriseServerReleases.latest}/user/articles/set-up-git`)
+      const res = await get(
+        `/en/enterprise/${enterpriseServerReleases.latest}/user/articles/fork-a-repo`,
+      )
       expect(res.statusCode).toBe(301)
       expect(res.headers.location).toBe(userArticle)
     })
@@ -291,38 +327,37 @@ describe('redirects', () => {
     })
 
     test('no language code redirects to english', async () => {
-      const res = await get(`/enterprise/${enterpriseServerReleases.latest}/user/github/getting-started-with-github/set-up-git`)
-      expect(res.statusCode).toBe(301)
+      const res = await get(
+        `/enterprise/${enterpriseServerReleases.latest}/user/articles/fork-a-repo`,
+      )
+      expect(res.statusCode).toBe(302)
       expect(res.headers.location).toBe(userArticle)
     })
 
     test('no version redirects to latest version', async () => {
-      const res = await get('/en/enterprise/user/github/getting-started-with-github/set-up-git')
+      const res = await get('/en/enterprise/articles/fork-a-repo')
       expect(res.statusCode).toBe(301)
       expect(res.headers.location).toBe(userArticle)
-    })
-
-    test('no version redirects to latest version (japanese)', async () => {
-      const res = await get('/ja/enterprise/user/github/getting-started-with-github/set-up-git')
-      expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe(japaneseUserArticle)
     })
   })
 
   describe('enterprise user article with frontmatter redirect', () => {
-    const userArticle = `/en/enterprise-server@${enterpriseServerReleases.latest}/github/getting-started-with-github/access-permissions-on-github`
-    const redirectFromPath = '/articles/what-are-the-different-access-permissions'
-    const japaneseUserArticle = userArticle.replace('/en/', '/ja/')
+    const userArticle = `/en/enterprise-server@${enterpriseServerReleases.latest}/get-started/quickstart/fork-a-repo`
+    const redirectFromPath = '/articles/fork-a-repo'
 
     test('redirects to expected article', async () => {
-      const res = await get(`/en/enterprise/${enterpriseServerReleases.latest}/user${redirectFromPath}`)
+      const res = await get(
+        `/en/enterprise/${enterpriseServerReleases.latest}/user${redirectFromPath}`,
+      )
       expect(res.statusCode).toBe(301)
       expect(res.headers.location).toBe(userArticle)
     })
 
     test('no language code redirects to english', async () => {
-      const res = await get(`/enterprise/${enterpriseServerReleases.latest}/user${redirectFromPath}`)
-      expect(res.statusCode).toBe(301)
+      const res = await get(
+        `/enterprise/${enterpriseServerReleases.latest}/user${redirectFromPath}`,
+      )
+      expect(res.statusCode).toBe(302)
       expect(res.headers.location).toBe(userArticle)
     })
 
@@ -331,34 +366,73 @@ describe('redirects', () => {
       expect(res.statusCode).toBe(301)
       expect(res.headers.location).toBe(userArticle)
     })
-
-    test('no version redirects to latest version (japanese)', async () => {
-      const res = await get(`/ja/enterprise/user${redirectFromPath}`)
-      expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe(japaneseUserArticle)
-    })
   })
 
   describe('desktop guide', () => {
-    const desktopGuide = '/en/desktop/contributing-and-collaborating-using-github-desktop/creating-an-issue-or-pull-request'
-    const japaneseDesktopGuides = desktopGuide.replace('/en/', '/ja/')
+    const desktopGuide =
+      '/en/desktop/working-with-your-remote-repository-on-github-or-github-enterprise/creating-an-issue-or-pull-request-from-github-desktop'
 
     test('no language code redirects to english', async () => {
-      const res = await get('/desktop/contributing-and-collaborating-using-github-desktop/creating-an-issue-or-pull-request')
-      expect(res.statusCode).toBe(301)
+      const res = await get(
+        '/desktop/contributing-and-collaborating-using-github-desktop/creating-an-issue-or-pull-request',
+      )
+      expect(res.statusCode).toBe(302)
       expect(res.headers.location).toBe(desktopGuide)
     })
 
     test('desktop/guides redirects to desktop', async () => {
-      const res = await get('/en/desktop/guides/contributing-and-collaborating-using-github-desktop/creating-an-issue-or-pull-request')
+      const res = await get(
+        '/en/desktop/guides/contributing-and-collaborating-using-github-desktop/creating-an-issue-or-pull-request',
+      )
       expect(res.statusCode).toBe(301)
       expect(res.headers.location).toBe(desktopGuide)
     })
+  })
 
-    test('desktop/guides redirects to desktop (japanese)', async () => {
-      const res = await get('/ja/desktop/guides/contributing-and-collaborating-using-github-desktop/creating-an-issue-or-pull-request')
+  describe('recently deprecated ghes version redirects that lack language', () => {
+    test('test redirect to an active enterprise-server version', async () => {
+      const url = `/enterprise-server@${enterpriseServerReleases.latest}/admin/release-notes`
+      const res = await get(url)
+      expect(res.statusCode).toBe(302)
+      expect(res.headers.location).toBe(`/en${url}`)
+    })
+    test('test redirect to a deprecated old enterprise-server version', async () => {
+      const url = `/enterprise-server@2.22/admin/release-notes`
+      const res = await get(url)
+      expect(res.statusCode).toBe(302)
+      expect(res.headers.location).toBe(`/en${url}`)
+    })
+    test('test redirect to a recently deprecated enterprise-server version', async () => {
+      const url = `/enterprise-server@${deprecatedWithFunctionalRedirects[0]}/admin/release-notes`
+      const res = await get(url)
+      expect(res.statusCode).toBe(302)
+      expect(res.headers.location).toBe(`/en${url}`)
+    })
+    test('any enterprise-server deprecated with functional redirects', async () => {
+      const url = `/enterprise-server@${deprecatedWithFunctionalRedirects[0]}`
+      const res = await get(url)
+      expect(res.statusCode).toBe(302)
+      expect(res.headers.location).toBe(`/en${url}`)
+    })
+  })
+
+  describe('redirects with double-slashes', () => {
+    test('prefix double-slash', async () => {
+      const res = await get(`//en`)
       expect(res.statusCode).toBe(301)
-      expect(res.headers.location).toBe(japaneseDesktopGuides)
+      expect(res.headers.location).toBe(`/en`)
+    })
+
+    test('double-slash elsewhere in the URL', async () => {
+      const res = await get(`/en//rest`)
+      expect(res.statusCode).toBe(301)
+      expect(res.headers.location).toBe(`/en/rest`)
+    })
+
+    test('double-slash trailing in the URL', async () => {
+      const res = await get(`/en//`)
+      expect(res.statusCode).toBe(301)
+      expect(res.headers.location).toBe(`/en`)
     })
   })
 })
