@@ -1,4 +1,4 @@
-const filenameToKey = require('../../../lib/filename-to-key')
+import filenameToKey from '../../../lib/filename-to-key.js'
 
 describe('filename-to-key', () => {
   test('converts filenames to object keys', () => {
