@@ -1,68 +1,56 @@
-const path = require('path')
-const cheerio = require('cheerio')
-const Page = require('../../lib/page')
-const prerenderedObjects = require('../../lib/graphql/static/prerendered-objects')
-const allVersions = require('../../lib/all-versions')
-const enterpriseServerReleases = require('../../lib/enterprise-server-releases')
-const nonEnterpriseDefaultVersion = require('../../lib/non-enterprise-default-version')
-const { latest } = require('../../lib/enterprise-server-releases')
+import { fileURLToPath } from 'url'
+import path from 'path'
+import cheerio from 'cheerio'
+import { describe, expect } from '@jest/globals'
 
-const getLinkData = require('../../lib/get-link-data')
-jest.mock('../../lib/get-link-data')
+import Page, { FrontmatterErrorsError } from '../../lib/page.js'
+import { allVersions } from '#src/versions/lib/all-versions.js'
+import enterpriseServerReleases, { latest } from '#src/versions/lib/enterprise-server-releases.js'
+import nonEnterpriseDefaultVersion from '#src/versions/lib/non-enterprise-default-version.js'
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
+const enterpriseServerVersions = Object.keys(allVersions).filter((v) =>
+  v.startsWith('enterprise-server@'),
+)
 
 // get the `free-pro-team` segment of `free-pro-team@latest`
 const nonEnterpriseDefaultPlan = nonEnterpriseDefaultVersion.split('@')[0]
 
 const opts = {
-  relativePath: 'github/collaborating-with-issues-and-pull-requests/about-branches.md',
+  relativePath:
+    'pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-branches.md',
   basePath: path.join(__dirname, '../../content'),
-  languageCode: 'en'
+  languageCode: 'en',
 }
 
 describe('Page class', () => {
   test('preserves file path info', async () => {
     const page = await Page.init(opts)
-    expect(page.relativePath).toBe('github/collaborating-with-issues-and-pull-requests/about-branches.md')
+    expect(page.relativePath).toBe(opts.relativePath)
     expect(page.fullPath.includes(page.relativePath)).toBe(true)
   })
 
-  test('does not error out on translated TOC with no links', async () => {
-    const page = await Page.init({
-      relativePath: 'translated-toc-with-no-links-index.md',
-      basePath: path.join(__dirname, '../fixtures'),
-      languageCode: 'ja'
-    })
-    expect(typeof page.title).toBe('string')
-  })
-
   describe('showMiniToc page property', () => {
-    let article, articleWithFM, tocPage, mapTopic
+    let article, articleWithFM, tocPage
 
     beforeAll(async () => {
       article = await Page.init({
         relativePath: 'sample-article.md',
         basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
 
       articleWithFM = await Page.init({
         showMiniToc: false,
         relativePath: article.relativePath,
         basePath: article.basePath,
-        languageCode: article.languageCode
+        languageCode: article.languageCode,
       })
 
       tocPage = await Page.init({
         relativePath: 'sample-toc-index.md',
         basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
-      })
-
-      mapTopic = await Page.init({
-        mapTopic: true,
-        relativePath: article.relativePath,
-        basePath: article.basePath,
-        languageCode: article.languageCode
+        languageCode: 'en',
       })
     })
 
@@ -74,87 +62,13 @@ describe('Page class', () => {
       expect(articleWithFM.showMiniToc).toBe(false)
     })
 
+    // products, categories, and map topics have index.md pages
     test('is undefined by default on index.md pages', () => {
       expect(tocPage.showMiniToc).toBeUndefined()
-    })
-
-    test('is undefined by default on map topics', () => {
-      expect(mapTopic.showMiniToc).toBeUndefined()
     })
   })
 
   describe('page.render(context)', () => {
-    test('rewrites links to include the current language prefix and version', async () => {
-      const page = await Page.init(opts)
-      const context = {
-        page: { version: `enterprise-server@${enterpriseServerReleases.latest}` },
-        currentVersion: `enterprise-server@${enterpriseServerReleases.latest}`,
-        currentPath: '/en/github/collaborating-with-issues-and-pull-requests/about-branches',
-        currentLanguage: 'en'
-      }
-      const rendered = await page.render(context)
-      const $ = cheerio.load(rendered)
-      expect(page.markdown.includes('(/articles/about-pull-requests)')).toBe(true)
-      expect(page.markdown.includes('(/en/articles/about-pull-requests)')).toBe(false)
-      expect($('a[href="/articles/about-pull-requests"]').length).toBe(0)
-      expect($(`a[href="/en/${`enterprise-server@${enterpriseServerReleases.latest}`}/articles/about-pull-requests"]`).length).toBeGreaterThan(0)
-    })
-
-    test('rewrites links on prerendered GraphQL page include the current language prefix and version', async () => {
-      const graphqlVersion = allVersions[`enterprise-server@${enterpriseServerReleases.latest}`].miscVersionName
-      const $ = cheerio.load(prerenderedObjects[graphqlVersion].html)
-      expect($('a[href^="/graphql/reference/input-objects"]').length).toBe(0)
-      expect($(`a[href^="/en/enterprise-server@${enterpriseServerReleases.latest}/graphql/reference/input-objects"]`).length).toBeGreaterThan(0)
-    })
-
-    test('rewrites links in the intro to include the current language prefix and version', async () => {
-      const page = await Page.init(opts)
-      page.rawIntro = '[Pull requests](/articles/about-pull-requests)'
-      const context = {
-        page: { version: nonEnterpriseDefaultVersion },
-        currentVersion: nonEnterpriseDefaultVersion,
-        currentPath: '/en/github/collaborating-with-issues-and-pull-requests/about-branches',
-        currentLanguage: 'en'
-      }
-      await page.render(context)
-      const $ = cheerio.load(page.intro)
-      expect($('a[href="/articles/about-pull-requests"]').length).toBe(0)
-      expect($('a[href="/en/articles/about-pull-requests"]').length).toBeGreaterThan(0)
-    })
-
-    test('does not rewrite links that include deprecated enterprise release numbers', async () => {
-      const page = await Page.init({
-        relativePath: 'admin/enterprise-management/migrating-from-github-enterprise-1110x-to-2123.md',
-        basePath: path.join(__dirname, '../../content'),
-        languageCode: 'en'
-      })
-      const context = {
-        page: { version: `enterprise-server@${enterpriseServerReleases.latest}` },
-        currentVersion: `enterprise-server@${enterpriseServerReleases.latest}`,
-        currentPath: `/en/enterprise-server@${enterpriseServerReleases.latest}/admin/enterprise-management/migrating-from-github-enterprise-1110x-to-2123`,
-        currentLanguage: 'en'
-      }
-      const rendered = await page.render(context)
-      const $ = cheerio.load(rendered)
-      expect(page.markdown.includes('(/enterprise/11.10.340/admin/articles/upgrading-to-the-latest-release/)')).toBe(true)
-      expect($(`a[href="/en/enterprise-server@${enterpriseServerReleases.latest}/11.10.340/admin/articles/upgrading-to-the-latest-release"]`).length).toBe(0)
-      expect($('a[href="/en/enterprise/11.10.340/admin/articles/upgrading-to-the-latest-release"]').length).toBeGreaterThan(0)
-    })
-
-    test('does not rewrite links to external redirects', async () => {
-      const page = await Page.init(opts)
-      page.markdown = `${page.markdown}\n\nSee [Capistrano](/capistrano).`
-      const context = {
-        page: { version: nonEnterpriseDefaultVersion },
-        currentVersion: nonEnterpriseDefaultVersion,
-        currentPath: `/en/${nonEnterpriseDefaultVersion}/github/collaborating-with-issues-and-pull-requests/about-branches`,
-        currentLanguage: 'en'
-      }
-      const rendered = await page.render(context)
-      const $ = cheerio.load(rendered)
-      expect($('a[href="/capistrano"]').length).toBe(1)
-    })
-
     // Most of our Liquid versioning tests are in https://github.com/docs/render-content,
     // But they don't have access to our currently supported versions, which we're testing here.
     // This test ensures that this works as expected: {% if enterpriseServerVersions contains currentVersion %}
@@ -162,33 +76,42 @@ describe('Page class', () => {
       const page = await Page.init({
         relativePath: 'page-versioned-for-all-enterprise-releases.md',
         basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
       // set version to the latest enterprise version
       const context = {
         currentVersion: `enterprise-server@${enterpriseServerReleases.latest}`,
         currentLanguage: 'en',
-        enterpriseServerVersions: Object.keys(allVersions).filter(id => id.startsWith('enterprise-server@'))
+        enterpriseServerVersions,
       }
+      context.currentPath = `/${context.currentLanguage}/${context.currentVersion}/${page.relativePath}`
       let rendered = await page.render(context)
       let $ = cheerio.load(rendered)
-      expect($.text()).toBe('This text should render on any actively supported version of Enterprise Server')
+      expect($.text()).toBe(
+        'This text should render on any actively supported version of Enterprise Server',
+      )
       expect($.text()).not.toBe('This text should only render on non-Enterprise')
 
       // change version to the oldest enterprise version, re-render, and test again;
       // the results should be the same
       context.currentVersion = `enterprise-server@${enterpriseServerReleases.oldestSupported}`
+      context.currentPath = `/${context.currentLanguage}/${context.currentVersion}/${page.relativePath}`
       rendered = await page.render(context)
       $ = cheerio.load(rendered)
-      expect($.text()).toBe('This text should render on any actively supported version of Enterprise Server')
+      expect($.text()).toBe(
+        'This text should render on any actively supported version of Enterprise Server',
+      )
       expect($.text()).not.toBe('This text should only render on non-Enterprise')
 
       // change version to non-enterprise, re-render, and test again;
       // the results should be the opposite
       context.currentVersion = nonEnterpriseDefaultVersion
+      context.currentPath = `/${context.currentLanguage}/${context.currentVersion}/${page.relativePath}`
       rendered = await page.render(context)
       $ = cheerio.load(rendered)
-      expect($.text()).not.toBe('This text should render on any actively supported version of Enterprise Server')
+      expect($.text()).not.toBe(
+        'This text should render on any actively supported version of Enterprise Server',
+      )
       expect($.text()).toBe('This text should only render on non-Enterprise')
     })
 
@@ -197,29 +120,16 @@ describe('Page class', () => {
       const page = await Page.init({
         relativePath: 'page-versioned-for-next-enterprise-release.md',
         basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
       // set version to 3.0
       const context = {
         currentVersion: 'enterprise-server@3.0',
-        currentLanguage: 'en'
+        currentLanguage: 'en',
       }
-      await expect(() => { return page.render(context) }).not.toThrow()
-    })
-
-    test('support next GitHub AE version in frontmatter', async () => {
-      // This fixture has `github-ae: 'next'` hardcoded in the frontmatter
-      const page = await Page.init({
-        relativePath: 'page-versioned-for-ghae-next.md',
-        basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
-      })
-      // set version to @latest
-      const context = {
-        currentVersion: 'github-ae@latest',
-        currentLanguage: 'en'
-      }
-      await expect(() => { return page.render(context) }).not.toThrow()
+      await expect(() => {
+        return page.render(context)
+      }).not.toThrow()
     })
   })
 
@@ -232,21 +142,21 @@ describe('Page class', () => {
     let page = await Page.init({
       relativePath: 'github/some-category/some-article.md',
       basePath: path.join(__dirname, '../fixtures/products'),
-      languageCode: 'en'
+      languageCode: 'en',
     })
     expect(page.parentProductId).toBe('github')
 
     page = await Page.init({
       relativePath: 'actions/some-category/some-article.md',
       basePath: path.join(__dirname, '../fixtures/products'),
-      languageCode: 'en'
+      languageCode: 'en',
     })
     expect(page.parentProductId).toBe('actions')
 
     page = await Page.init({
       relativePath: 'admin/some-category/some-article.md',
       basePath: path.join(__dirname, '../fixtures/products'),
-      languageCode: 'en'
+      languageCode: 'en',
     })
     expect(page.parentProductId).toBe('admin')
   })
@@ -259,45 +169,70 @@ describe('Page class', () => {
 
     test('has a key for every supported enterprise version (and no deprecated versions)', async () => {
       const page = await Page.init(opts)
-      const pageVersions = page.permalinks.map(permalink => permalink.pageVersion)
-      expect(enterpriseServerReleases.supported.every(version => pageVersions.includes(`enterprise-server@${version}`))).toBe(true)
-      expect(enterpriseServerReleases.deprecated.every(version => !pageVersions.includes(`enterprise-server@${version}`))).toBe(true)
+      const pageVersions = page.permalinks.map((permalink) => permalink.pageVersion)
+      expect(
+        enterpriseServerReleases.supported.every((version) =>
+          pageVersions.includes(`enterprise-server@${version}`),
+        ),
+      ).toBe(true)
+      expect(
+        enterpriseServerReleases.deprecated.every(
+          (version) => !pageVersions.includes(`enterprise-server@${version}`),
+        ),
+      ).toBe(true)
     })
 
     test('sets versioned values', async () => {
       const page = await Page.init(opts)
-      expect(page.permalinks.find(permalink => permalink.pageVersion === nonEnterpriseDefaultVersion).href).toBe('/en/github/collaborating-with-issues-and-pull-requests/about-branches')
-      expect(page.permalinks.find(permalink => permalink.pageVersion === `enterprise-server@${enterpriseServerReleases.oldestSupported}`).href).toBe(`/en/enterprise-server@${enterpriseServerReleases.oldestSupported}/github/collaborating-with-issues-and-pull-requests/about-branches`)
+      const expectedPath =
+        'pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-branches'
+      expect(
+        page.permalinks.find((permalink) => permalink.pageVersion === nonEnterpriseDefaultVersion)
+          .href,
+      ).toBe(`/en/${expectedPath}`)
+      expect(
+        page.permalinks.find(
+          (permalink) =>
+            permalink.pageVersion ===
+            `enterprise-server@${enterpriseServerReleases.oldestSupported}`,
+        ).href,
+      ).toBe(`/en/enterprise-server@${enterpriseServerReleases.oldestSupported}/${expectedPath}`)
     })
 
     test('homepage permalinks', async () => {
       const page = await Page.init({
         relativePath: 'index.md',
         basePath: path.join(__dirname, '../../content'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
-      expect(page.permalinks.find(permalink => permalink.pageVersion === 'homepage').href).toBe('/en')
-      expect(page.permalinks.find(permalink => permalink.pageVersion === `enterprise-server@${enterpriseServerReleases.oldestSupported}`).href).toBe(`/en/enterprise-server@${enterpriseServerReleases.oldestSupported}`)
-    })
-
-    test('permalinks for dotcom-only pages', async () => {
-      const page = await Page.init({
-        relativePath: 'github/getting-started-with-github/signing-up-for-a-new-github-account.md',
-        basePath: path.join(__dirname, '../../content'),
-        languageCode: 'en'
-      })
-      expect(page.permalinks.find(permalink => permalink.pageVersion === nonEnterpriseDefaultVersion).href).toBe('/en/github/getting-started-with-github/signing-up-for-a-new-github-account')
-      expect(page.permalinks.length).toBe(1)
+      expect(
+        page.permalinks.find((permalink) => permalink.pageVersion === nonEnterpriseDefaultVersion)
+          .href,
+      ).toBe('/en')
+      expect(
+        page.permalinks.find(
+          (permalink) =>
+            permalink.pageVersion ===
+            `enterprise-server@${enterpriseServerReleases.oldestSupported}`,
+        ).href,
+      ).toBe(`/en/enterprise-server@${enterpriseServerReleases.oldestSupported}`)
     })
 
     test('permalinks for enterprise-only pages', async () => {
       const page = await Page.init({
         relativePath: 'products/admin/some-category/some-article.md',
         basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
-      expect(page.permalinks.find(permalink => permalink.pageVersion === `enterprise-server@${enterpriseServerReleases.latest}`).href).toBe(`/en/enterprise-server@${enterpriseServerReleases.latest}/products/admin/some-category/some-article`)
-      const pageVersions = page.permalinks.map(permalink => permalink.pageVersion)
+      expect(
+        page.permalinks.find(
+          (permalink) =>
+            permalink.pageVersion === `enterprise-server@${enterpriseServerReleases.latest}`,
+        ).href,
+      ).toBe(
+        `/en/enterprise-server@${enterpriseServerReleases.latest}/products/admin/some-category/some-article`,
+      )
+      const pageVersions = page.permalinks.map((permalink) => permalink.pageVersion)
       expect(pageVersions.length).toBeGreaterThan(1)
       expect(pageVersions.includes(nonEnterpriseDefaultVersion)).toBe(false)
     })
@@ -306,134 +241,59 @@ describe('Page class', () => {
       const page = await Page.init({
         relativePath: 'products/actions/some-category/some-article.md',
         basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
-      expect(page.permalinks.find(permalink => permalink.pageVersion === nonEnterpriseDefaultVersion).href).toBe('/en/products/actions/some-category/some-article')
+      expect(
+        page.permalinks.find((permalink) => permalink.pageVersion === nonEnterpriseDefaultVersion)
+          .href,
+      ).toBe('/en/products/actions/some-category/some-article')
       expect(page.permalinks.length).toBe(1)
     })
-
-    test('permalinks for non-GitHub.com products with Enterprise versions', async () => {
-      const page = await Page.init({
-        relativePath: '/insights/installing-and-configuring-github-insights/about-github-insights.md',
-        basePath: path.join(__dirname, '../../content'),
-        languageCode: 'en'
-      })
-      expect(page.permalinks.find(permalink => permalink.pageVersion === `enterprise-server@${enterpriseServerReleases.latest}`).href).toBe(`/en/enterprise-server@${enterpriseServerReleases.latest}/insights/installing-and-configuring-github-insights/about-github-insights`)
-      const pageVersions = page.permalinks.map(permalink => permalink.pageVersion)
-      expect(pageVersions.length).toBeGreaterThan(1)
-      expect(pageVersions.includes(nonEnterpriseDefaultVersion)).toBe(false)
-    })
   })
 
-  describe('learning tracks', () => {
+  describe('videos', () => {
     let page
 
     beforeEach(async () => {
       page = await Page.init({
-        relativePath: 'article-with-learning-tracks.md',
+        relativePath: 'article-with-videos.md',
         basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
     })
 
-    it('includes learning tracks specified in frontmatter', async () => {
-      expect(page.learningTracks).toStrictEqual([
-        'track_1',
-        'track_2',
-        'non_existing_track',
-        '{% if currentVersion == "free-pro-team@latest" %}dotcom_only_track{% endif %}',
-        '{% if currentVersion != "free-pro-team@latest" %}enterprise_only_track{% endif %}'
+    it('includes videos specified in the featuredLinks frontmatter', async () => {
+      expect(page.featuredLinks.videos).toStrictEqual([
+        {
+          title: 'codespaces',
+          href: 'https://www.youtube-nocookie.com/embed/_W9B7qc9lVc',
+        },
+        {
+          title: 'more codespaces',
+          href: 'https://www.youtube-nocookie.com/embed/_W9B7qc9lVc',
+        },
+        {
+          title: 'even more codespaces',
+          href: 'https://www.youtube-nocookie.com/embed/_W9B7qc9lVc',
+        },
       ])
-    })
 
-    it('renders learning tracks that have been defined', async () => {
-      const guides = ['/path/guide1', '/path/guide2']
-      const context = {
-        currentLanguage: 'en',
-        currentProduct: 'snowbird',
-        currentVersion: nonEnterpriseDefaultVersion,
-        site: {
-          data: {
-            'learning-tracks': {
-              snowbird: {
-                track_1: {
-                  title: 'title',
-                  description: 'description',
-                  guides
-                },
-                track_2: {
-                  title: 'title',
-                  description: 'description',
-                  guides
-                },
-                dotcom_only_track: {
-                  title: 'title',
-                  description: 'description',
-                  guides
-                },
-                enterprise_only_track: {
-                  title: 'title',
-                  description: 'description',
-                  guides
-                }
-              }
-            }
-          }
-        }
-      }
-      await page.render(context)
-      // To actually render the guides in this test, we would have to load context.pages and context.redirects;
-      // To avoid that we can just test that the function was called with the expected data.
-      expect(getLinkData).toHaveBeenCalledWith(guides, context)
-      expect(page.learningTracks).toHaveLength(3)
-      expect(page.learningTracks.find(track => track.trackName === 'dotcom_only_track')).toBeTruthy()
-
-      // Test that Liquid versioning is respected
-      context.currentVersion = `enterprise-server@${latest}`
-      await page.render(context)
-      expect(page.learningTracks).toHaveLength(3)
-      expect(page.learningTracks.find(track => track.trackName === 'dotcom_only_track')).toBeFalsy()
-      expect(page.learningTracks.find(track => track.trackName === 'enterprise_only_track')).toBeTruthy()
+      expect(page.featuredLinks.videosHeading).toBe('Custom Videos heading')
     })
   })
 
-  describe('includeGuides', () => {
-    let page
-
-    beforeEach(async () => {
-      page = await Page.init({
-        relativePath: 'article-with-includeGuides.md',
+  describe('introLinks', () => {
+    it('includes the links specified in the introLinks frontmatter', async () => {
+      const page = await Page.init({
+        relativePath: 'article-with-introLinks.md',
         basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
-    })
 
-    it('includes guide paths specified in frontmatter', async () => {
-      expect(page.includeGuides).toStrictEqual(['/path/guide1', '/path/guide2', '/path/guide3'])
-    })
-
-    it('renders guides and topics', async () => {
-      getLinkData.mockImplementation(() => {
-        return [{
-          page: { topics: ['Spring', 'Summer'] }
-        }, {
-          page: { topics: ['Summer', 'Fall'] }
-        }, {
-          page: { topics: ['Fall', 'Winter'] }
-        }]
+      expect(page.introLinks).toStrictEqual({
+        overview: 'https://github.com',
+        'custom link!': 'https://github.com/features',
       })
-      const guides = ['/path/guide1', '/path/guide2', '/path/guide3']
-      const context = {
-        currentVersion: nonEnterpriseDefaultVersion,
-        currentLanguage: 'en'
-      }
-      await page.render(context)
-      expect(getLinkData).toHaveBeenCalledWith(guides, context)
-      expect(page.includeGuides).toHaveLength(3)
-      expect(page.allTopics).toHaveLength(4)
-      expect(page.allTopics).toEqual(
-        expect.arrayContaining(['Spring', 'Summer', 'Fall', 'Winter'])
-      )
     })
   })
 
@@ -446,105 +306,139 @@ describe('Page class', () => {
     })
   })
 
-  describe('Page.getLanguageVariants()', () => {
-    it('returns an array of language variants of the given URL', () => {
-      const variants = Page.getLanguageVariants('/en')
-      expect(variants.every(variant => variant.name)).toBe(true)
-      expect(variants.every(variant => variant.code)).toBe(true)
-      expect(variants.every(variant => variant.href)).toBe(true)
-    })
-
-    it('works for the homepage', () => {
-      const variants = Page.getLanguageVariants('/en')
-      expect(variants.find(({ code }) => code === 'en').href).toBe('/en')
-      // expect(variants.find(({ code }) => code === 'ja').href).toBe('/ja')
-    })
-
-    it('works for enterprise URLs', () => {
-      const variants = Page.getLanguageVariants(`/ja/enterprise/${enterpriseServerReleases.oldestSupported}/user/articles/github-glossary`)
-      expect(variants.find(({ code }) => code === 'en').href).toBe(`/en/enterprise/${enterpriseServerReleases.oldestSupported}/user/articles/github-glossary`)
-      // expect(variants.find(({ code }) => code === 'ja').href).toBe('/ja/enterprise/2.14/user/articles/github-glossary')
-    })
-  })
-
-  test('fixes translated frontmatter that includes verdadero', async () => {
-    const page = await Page.init({
-      relativePath: 'article-with-mislocalized-frontmatter.md',
-      basePath: path.join(__dirname, '../fixtures'),
-      languageCode: 'ja'
-    })
-    expect(page.mapTopic).toBe(true)
-  })
-
   describe('page.versions frontmatter', () => {
-    test.skip('pages that apply to older enterprise versions', async () => {
-    // There are none of these in the content at this time!
-    })
-
-    // Note this test will go out of date when we deprecate 2.20
-    test('pages that apply to newer enterprise versions', async () => {
+    test('pages that use short names in versions frontmatter', async () => {
       const page = await Page.init({
-        relativePath: 'github/administering-a-repository/comparing-releases.md',
-        basePath: path.join(__dirname, '../../content'),
-        languageCode: 'en'
+        relativePath: 'short-versions.md',
+        basePath: path.join(__dirname, '../fixtures'),
+        languageCode: 'en',
       })
-      expect(nonEnterpriseDefaultPlan in page.versions).toBe(true)
-      expect(page.versions['enterprise-server']).toBe('>=2.21')
+      expect(page.versions.fpt).toBe('*')
+      expect(page.versions.ghes).toBe('>3.0')
+      expect(page.versions.ghae).toBeUndefined()
+      expect(page.applicableVersions.includes('free-pro-team@latest')).toBe(true)
+      expect(page.applicableVersions.includes(`enterprise-server@${latest}`)).toBe(true)
+      expect(page.applicableVersions.includes(`github-ae@latest`)).toBe(false)
     })
 
     test('index page', async () => {
       const page = await Page.init({
         relativePath: 'index.md',
         basePath: path.join(__dirname, '../../content'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
-      expect(page.versions).toBe('*')
+      expect(page.versions).toEqual({ fpt: '*', ghae: '*', ghec: '*', ghes: '*' })
     })
 
     test('enterprise admin index page', async () => {
       const page = await Page.init({
         relativePath: 'admin/index.md',
         basePath: path.join(__dirname, '../../content'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
 
       expect(nonEnterpriseDefaultPlan in page.versions).toBe(false)
-      expect(page.versions['enterprise-server']).toBe('*')
+      expect(page.versions.ghes).toBe('*')
+    })
+
+    test('feature versions frontmatter', async () => {
+      // This fixture file has the frontmatter:
+      //
+      // versions:
+      //   fpt: '*'
+      //   ghes: '*'
+      //   feature: 'placeholder'
+      //
+      // and placeholder.yml has:
+      //
+      // versions:
+      //   ghes: '<3.0'
+      //   ghae: '*'
+      //
+      // So we expect to get the versioning from both.
+      const page = await Page.init({
+        relativePath: 'feature-versions-frontmatter.md',
+        basePath: path.join(__dirname, '../fixtures'),
+        languageCode: 'en',
+      })
+
+      // Test the raw page data.
+      expect(page.versions.fpt).toBe('*')
+      expect(page.versions.ghes).toBe('>2.21')
+      expect(page.versions.ghae).toBeUndefined()
+
+      // Test the resolved versioning, where GHES releases specified in frontmatter and in
+      // feature versions are combined (i.e., one doesn't overwrite the other).
+      // We can't test that GHES 2.21 is _not_ included here (which it shouldn't be),
+      // because lib/get-applicable-versions only returns currently supported versions,
+      // so as soon as 2.21 is deprecated, a test for that _not_ to exist will not be meaningful.
+      // But by testing that the _latest_ GHES version is returned, we can ensure that the
+      // the frontmatter GHES `*` is not being overwritten by the placeholder's GHES `<3.0`.
+      expect(page.applicableVersions.includes('free-pro-team@latest')).toBe(true)
+      expect(page.applicableVersions.includes(`enterprise-server@${latest}`)).toBe(true)
+      expect(page.applicableVersions.includes('github-ae@latest')).toBe(true)
+      expect(page.applicableVersions.includes('feature')).toBe(false)
+      expect(page.applicableVersions.includes('placeholder')).toBe(false)
     })
   })
 
   describe('platform specific content', () => {
     test('page.defaultPlatform frontmatter', async () => {
       const page = await Page.init({
-        relativePath: 'actions/hosting-your-own-runners/configuring-the-self-hosted-runner-application-as-a-service.md',
-        basePath: path.join(__dirname, '../../content'),
-        languageCode: 'en'
+        relativePath: 'actions/some-category/some-article.md',
+        basePath: path.join(__dirname, '../fixtures/products'),
+        languageCode: 'en',
       })
       expect(page.defaultPlatform).toBeDefined()
       expect(page.defaultPlatform).toBe('linux')
+    })
+  })
+
+  describe('tool specific content', () => {
+    test('page.defaultTool frontmatter', async () => {
+      const page = await Page.init({
+        relativePath: 'default-tool.md',
+        basePath: path.join(__dirname, '../fixtures'),
+        languageCode: 'en',
+      })
+      expect(page.defaultTool).toBeDefined()
+      expect(page.defaultTool).toBe('cli')
     })
   })
 })
 
 describe('catches errors thrown in Page class', () => {
   test('frontmatter parsing error', async () => {
-    async function getPage () {
+    async function getPage() {
       return await Page.init({
         relativePath: 'page-with-frontmatter-error.md',
         basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
     }
 
-    await expect(getPage).rejects.toThrowError('invalid frontmatter entry')
+    await expect(getPage).rejects.toThrow(FrontmatterErrorsError)
   })
 
   test('missing versions frontmatter', async () => {
-    async function getPage () {
+    async function getPage() {
       return await Page.init({
         relativePath: 'page-with-missing-product-versions.md',
         basePath: path.join(__dirname, '../fixtures'),
-        languageCode: 'en'
+        languageCode: 'en',
+      })
+    }
+
+    await expect(getPage).rejects.toThrowError('versions')
+  })
+
+  // TODO - UNSKIP WHEN GHAE IS UPDATED WITH SEMVER VERSIONING
+  test.skip('invalid versions frontmatter', async () => {
+    async function getPage() {
+      return await Page.init({
+        relativePath: 'page-with-invalid-product-version.md',
+        basePath: path.join(__dirname, '../fixtures'),
+        languageCode: 'en',
       })
     }
 
@@ -552,26 +446,46 @@ describe('catches errors thrown in Page class', () => {
   })
 
   test('English page with a version in frontmatter that its parent product is not available in', async () => {
-    async function getPage () {
+    async function getPage() {
       return await Page.init({
         relativePath: 'admin/some-category/some-article-with-mismatched-versions-frontmatter.md',
         basePath: path.join(__dirname, '../fixtures/products'),
-        languageCode: 'en'
+        languageCode: 'en',
       })
     }
 
     expect(getPage).rejects.toThrowError(/`versions` frontmatter.*? product is not available in/)
   })
 
-  test('non-English page with a version in frontmatter that its parent product is not available in', async () => {
-    async function getPage () {
-      return await Page.init({
-        relativePath: 'admin/some-category/some-article-with-mismatched-versions-frontmatter.md',
-        basePath: path.join(__dirname, '../fixtures/products'),
-        languageCode: 'es'
+  describe('versionining optional attributes', () => {
+    test("re-rendering set appropriate 'product', 'permissions', 'learningTracks'", async () => {
+      const page = await Page.init({
+        relativePath: 'page-with-optional-attributes.md',
+        basePath: path.join(__dirname, '../fixtures'),
+        languageCode: 'en',
       })
-    }
+      const context = {
+        page: { version: `enterprise-server@3.2` },
+        currentVersion: `enterprise-server@3.2`,
+        currentVersionObj: {},
+        currentProduct: 'snowbird',
+        currentLanguage: 'en',
+        currentPath: '/en/enterprise-server@3.2/optional/attributes',
+        fpt: false, // what the shortVersions contextualizer does
+      }
 
-    await expect(getPage).rejects.toThrowError(/`versions` frontmatter.*? product is not available in/)
+      await page.render(context)
+      expect(page.product).toBe('')
+      expect(page.permissions).toBe('')
+
+      // Change to FPT
+      context.page.version = nonEnterpriseDefaultVersion
+      context.version = nonEnterpriseDefaultVersion
+      context.currentPath = '/en/optional/attributes'
+      context.fpt = true
+      await page.render(context)
+      expect(page.product).toContain('FPT rulez!')
+      expect(page.permissions).toContain('FPT only!')
+    })
   })
 })
