@@ -1,4 +1,6 @@
-const parse = require('../../lib/read-frontmatter')
+import parse from '../../lib/read-frontmatter.js'
+import { schema as frontmatterSchema } from '../../lib/frontmatter.js'
+
 const filepath = 'path/to/file.md'
 const fixture1 = `---
 title: Hello, World
@@ -6,6 +8,13 @@ meaning_of_life: 42
 ---
 
 I am content.
+`
+const fixture2 = `---
+versions:
+  fpt: '*'
+  ghec: '*'
+  ghes: 'BAD_VERSION'
+---
 `
 
 describe('frontmatter', () => {
@@ -36,7 +45,7 @@ I am content.
       const expectedError = {
         filepath: 'path/to/file.md',
         message: 'YML parsing error!',
-        reason: 'invalid frontmatter entry'
+        reason: 'invalid frontmatter entry',
       }
       expect(errors[0]).toEqual(expectedError)
     })
@@ -54,7 +63,7 @@ I am content.
       const expectedError = {
         filepath: 'path/to/file.md',
         message: 'YML parsing error!',
-        reason: 'bad indentation of a mapping entry'
+        reason: 'bad indentation of a mapping entry',
       }
       expect(errors[0]).toEqual(expectedError)
     })
@@ -65,12 +74,12 @@ I am content.
       const schema = {
         properties: {
           title: {
-            type: 'string'
+            type: 'string',
           },
           meaning_of_life: {
-            type: 'number'
-          }
-        }
+            type: 'number',
+          },
+        },
       }
 
       const { data, content, errors } = parse(fixture1, { schema })
@@ -85,9 +94,9 @@ I am content.
         properties: {
           meaning_of_life: {
             type: 'number',
-            minimum: 50
-          }
-        }
+            minimum: 50,
+          },
+        },
       }
 
       const { data, content, errors } = parse(fixture1, { schema })
@@ -96,127 +105,70 @@ I am content.
       expect(content.trim()).toBe('I am content.')
       expect(errors.length).toBe(1)
       const expectedError = {
-        attribute: 'minimum',
-        property: 'meaning_of_life',
-        expected: 50,
-        actual: 42,
-        message: 'must be greater than or equal to 50'
+        instancePath: '/meaning_of_life',
+        schemaPath: '#/properties/meaning_of_life/minimum',
+        keyword: 'minimum',
+        params: {
+          comparison: '>=',
+          limit: 50,
+        },
+        message: 'must be >= 50',
       }
+      expect(errors[0]).toEqual(expectedError)
+    })
+
+    it('creates errors if versions frontmatter does not match semver format', () => {
+      const schema = { type: 'object', required: ['versions'], properties: {} }
+      schema.properties.versions = Object.assign({}, frontmatterSchema.properties.versions)
+
+      const { errors } = parse(fixture2, { schema })
+      const expectedError = {
+        instancePath: '/versions/ghes',
+        schemaPath: '#/properties/versions/properties/ghes/errorMessage',
+        keyword: 'errorMessage',
+        params: {
+          errors: [
+            {
+              instancePath: '/versions/ghes',
+              schemaPath: '#/properties/versions/properties/ghes/format',
+              keyword: 'format',
+              params: {
+                format: 'semver',
+              },
+              message: 'must match format "semver"',
+              emUsed: true,
+            },
+          ],
+        },
+        message: 'Must be a valid SemVer range: "BAD_VERSION"',
+      }
+
       expect(errors[0]).toEqual(expectedError)
     })
 
     it('creates errors if required frontmatter is not present', () => {
       const schema = {
+        type: 'object',
+        required: ['yet_another_key'],
         properties: {
           yet_another_key: {
             type: 'string',
-            required: true
-          }
-        }
+          },
+        },
       }
 
       const { errors } = parse(fixture1, { schema })
       expect(errors.length).toBe(1)
       const expectedError = {
-        attribute: 'required',
-        property: 'yet_another_key',
-        expected: true,
-        actual: undefined,
-        message: 'is required'
+        instancePath: '',
+        schemaPath: '#/required',
+        keyword: 'required',
+        params: {
+          missingProperty: 'yet_another_key',
+        },
+        message: "must have required property 'yet_another_key'",
       }
       expect(errors[0]).toEqual(expectedError)
-    })
-  })
-
-  describe('validateKeyNames', () => {
-    const schema = {
-      properties: {
-        age: {
-          type: 'number'
-        }
-      }
-    }
-
-    it('creates errors for undocumented keys if `validateKeyNames` is true', () => {
-      const { errors } = parse(fixture1, { schema, validateKeyNames: true, filepath })
-      expect(errors.length).toBe(2)
-      const expectedErrors = [
-        {
-          property: 'title',
-          message: 'not allowed. Allowed properties are: age',
-          filepath: 'path/to/file.md'
-        },
-        {
-          property: 'meaning_of_life',
-          message: 'not allowed. Allowed properties are: age',
-          filepath: 'path/to/file.md'
-        }
-      ]
-      expect(errors).toEqual(expectedErrors)
-    })
-
-    it('does not create errors for undocumented keys if `validateKeyNames` is false', () => {
-      const { errors } = parse(fixture1, { schema, validateKeyNames: false })
-      expect(errors.length).toBe(0)
-    })
-  })
-
-  describe('validateKeyOrder', () => {
-    it('creates errors if `validateKeyOrder` is true and keys are not in order', () => {
-      const schema = {
-        properties: {
-          meaning_of_life: {
-            type: 'number'
-          },
-          title: {
-            type: 'string'
-          }
-        }
-      }
-      const { errors } = parse(fixture1, { schema, validateKeyOrder: true, filepath })
-      const expectedErrors = [
-        {
-          property: 'keys',
-          message: 'keys must be in order. Current: title,meaning_of_life; Expected: meaning_of_life,title',
-          filepath: 'path/to/file.md'
-        }
-      ]
-      expect(errors).toEqual(expectedErrors)
-    })
-
-    it('does not create errors if `validateKeyOrder` is true and keys are in order', () => {
-      const schema = {
-        properties: {
-          title: {
-            type: 'string'
-          },
-          meaning_of_life: {
-            type: 'number'
-          }
-        }
-      }
-      const { errors } = parse(fixture1, { schema, validateKeyOrder: true })
-      expect(errors.length).toBe(0)
-    })
-
-    it('does not create errors if `validateKeyOrder` is true and expected keys are in order', () => {
-      const schema = {
-        properties: {
-          title: {
-            type: 'string',
-            required: true
-          },
-          yet_another_key: {
-            type: 'string'
-          },
-          meaning_of_life: {
-            type: 'number',
-            required: true
-          }
-        }
-      }
-      const { errors } = parse(fixture1, { schema, validateKeyOrder: true })
-      expect(errors.length).toBe(0)
     })
   })
 })
