@@ -1,22 +1,25 @@
-const { loadPages } = require('../../lib/pages')
-const renderContent = require('../../lib/render-content')
-const allVersions = Object.keys(require('../../lib/all-versions'))
+import { jest } from '@jest/globals'
+import { loadPageMap, loadPages } from '../../lib/page-data.js'
+import { renderContent } from '#src/content-render/index.js'
+import { allVersions } from '#src/versions/lib/all-versions.js'
 
 describe('toc links', () => {
   jest.setTimeout(3 * 60 * 1000)
 
   test('every toc link works without redirects', async () => {
-    const pages = await loadPages()
+    const pageList = await loadPages()
 
-    const englishIndexPages = pages
-      .filter(page => page.languageCode === 'en' && page.relativePath.endsWith('index.md'))
+    const englishIndexPages = pageList.filter(
+      (page) => page.languageCode === 'en' && page.relativePath.endsWith('index.md'),
+    )
+    const pages = await loadPageMap(pageList)
 
     const issues = []
 
-    for (const pageVersion of allVersions) {
+    for (const pageVersion of Object.keys(allVersions)) {
       for (const page of englishIndexPages) {
         // skip page if it doesn't have a permalink for the current product version
-        if (!page.permalinks.some(permalink => permalink.pageVersion === pageVersion)) continue
+        if (!page.permalinks.some((permalink) => permalink.pageVersion === pageVersion)) continue
 
         // build fake context object for rendering the page
         const context = {
@@ -24,7 +27,8 @@ describe('toc links', () => {
           pages,
           redirects: {},
           currentLanguage: 'en',
-          currentVersion: pageVersion
+          currentVersion: pageVersion,
+          currentVersionObj: allVersions[pageVersion],
         }
 
         // ensure all toc pages can render
@@ -34,7 +38,7 @@ describe('toc links', () => {
           issues.push({
             'TOC path': page.relativePath,
             error: err.message,
-            pageVersion
+            pageVersion,
           })
         }
       }
