@@ -1,23 +1,5 @@
 #!/usr/bin/env node
 
-const fs = require('fs')
-const path = require('path')
-const walk = require('walk-sync')
-const slugger = new (require('github-slugger'))()
-const entities = new (require('html-entities').XmlEntities)()
-const frontmatter = require('../lib/read-frontmatter')
-const { execSync } = require('child_process')
-const addRedirectToFrontmatter = require('../lib/redirects/add-redirect-to-frontmatter')
-
-const contentDir = path.join(process.cwd(), 'content')
-
-const contentFiles = walk(contentDir, { includeBasePath: true, directories: false })
-  .filter(file => {
-    return file.endsWith('.md') &&
-    !file.endsWith('index.md') &&
-    !file.includes('README')
-  })
-
 // [start-readme]
 //
 // An automated test checks for discrepancies between filenames and [autogenerated heading IDs](https://www.npmjs.com/package/remark-autolink-headings).
@@ -27,13 +9,32 @@ const contentFiles = walk(contentDir, { includeBasePath: true, directories: fals
 //
 // [end-readme]
 
+import fs from 'fs'
+import path from 'path'
+import walk from 'walk-sync'
+import GithubSlugger from 'github-slugger'
+import { decode } from 'html-entities'
+import frontmatter from '../lib/read-frontmatter.js'
+import { execSync } from 'child_process'
+import addRedirectToFrontmatter from '../src/redirects/scripts/helpers/add-redirect-to-frontmatter.js'
+
+const slugger = new GithubSlugger()
+
+const contentDir = path.join(process.cwd(), 'content')
+
+const contentFiles = walk(contentDir, { includeBasePath: true, directories: false }).filter(
+  (file) => {
+    return file.endsWith('.md') && !file.endsWith('index.md') && !file.includes('README')
+  },
+)
+
 // TODO fix path separators in the redirect
 if (process.platform.startsWith('win')) {
   console.log('This script cannot be run on Windows at this time! Exiting...')
   process.exit()
 }
 
-contentFiles.forEach(oldFullPath => {
+contentFiles.forEach((oldFullPath) => {
   const { data, content } = frontmatter(fs.readFileSync(oldFullPath, 'utf8'))
 
   // skip pages with frontmatter flag
@@ -43,7 +44,7 @@ contentFiles.forEach(oldFullPath => {
   // where title = Foo bar
   // and slug = foo-bar
   slugger.reset()
-  const slug = slugger.slug(entities.decode(data.title))
+  const slug = slugger.slug(decode(data.title))
 
   // get the basename of each file
   // where file = content/foo-bar.md
